Uses a genome object to find context and add it to the variant table

Uses a genome object to find context and add it to the variant table

Usage

add_flank_to_variants(
  musica,
  g,
  flank_start,
  flank_end,
  build_table = TRUE,
  overwrite = FALSE
)

Arguments

musica

Input samples

g

A BSgenome object indicating which genome reference the variants and their coordinates were derived from.

flank_start

Start of flank area to add, can be positive or negative

flank_end

End of flank area to add, can be positive or negative

build_table

Automatically build a table using the annotation and add

overwrite

Overwrite existing count table

Value

None it to the musica

Examples

data(musica_sbs96_tiny)
g <- select_genome("19")
add_flank_to_variants(musica_sbs96_tiny, g, 1, 2)
add_flank_to_variants(musica_sbs96_tiny, g, -2, -1)