Automatic filtering of signatures for exposure prediction gridded across specific annotation

Automatic filtering of signatures for exposure prediction gridded across specific annotation

Usage

auto_predict_grid(
  musica,
  modality,
  signature_res,
  algorithm,
  model_id = NULL,
  result_name = "result",
  sample_annotation = NULL,
  min_exists = 0.05,
  proportion_samples = 0.25,
  rare_exposure = 0.4,
  verbose = TRUE,
  combine_res = TRUE,
  make_copy = FALSE,
  table_name = NULL
)

Arguments

musica

Input samples to predict signature weights

modality

Modality used for posterior prediction (e.g. SBS96)

signature_res

Signatures to automatically subset from for prediction

algorithm

Algorithm to use for prediction. Choose from "lda_posterior", and decompTumor2Sig

model_id

Name of model

result_name

Name for result_list entry to save the results to. Default "result".

sample_annotation

Annotation to grid across, if none given, prediction subsetting on all samples together

min_exists

Threshold to consider a signature active in a sample

proportion_samples

Threshold of samples to consider a signature active in the cohort

rare_exposure

A sample will be considered active in the cohort if at least one sample has more than this threshold proportion

verbose

Print current annotation value being predicted on

combine_res

Automatically combines a list of annotation results into a single result object with zero exposure values for signatures not found in a given annotation's set of samples

make_copy

If FALSE, the inputted musica object is updated and nothing is returned. If TRUE, a new musica object is created and returned. Default FALSE.

table_name

Use modality instead

Value

Returns nothing or a new musica object, depending on the make_copy parameter.

Examples

data(musica_annot)
data(cosmic_v2_sigs)
auto_predict_grid(
  musica = musica_annot, modality = "SBS96",
  signature_res = cosmic_v2_sigs, algorithm = "lda",
  sample_annotation = "Tumor_Subtypes"
)
#> Lung
#> Breast
auto_predict_grid(musica_annot, "SBS96", cosmic_v2_sigs, "lda")