Automatic filtering of inactive signatures

Automatic filtering of inactive signatures

auto_subset_sigs(
  musica,
  table_name,
  signature_res,
  algorithm,
  min_exists = 0.05,
  proportion_samples = 0.25,
  rare_exposure = 0.4
)

Arguments

musica
A musica object.

table_name
Name of table used for posterior prediction (e.g. SBS96)

signature_res
Signatures to automatically subset from for prediction

algorithm
Algorithm to use for prediction. Choose from "lda_posterior", decompTumor2Sig, and deconstructSigs

min_exists
Threshold to consider a signature active in a sample

proportion_samples
Threshold of samples to consider a signature active in the cohort

rare_exposure
A sample will be considered active in the cohort if at least one sample has more than this threshold proportion

Value

A result object containing automatically subset signatures and corresponding sample weights