Combines tables into a single table that can be used for discovery/prediction

Combines tables into a single table that can be used for discovery/prediction

Usage

combine_count_tables(
  musica,
  to_comb,
  name,
  description = character(),
  color_variable = character(),
  color_mapping = character(),
  overwrite = FALSE
)

Arguments

musica

A musica object.

to_comb

A vector of table names to combine. Each table must already exist within the input musica object

name

Name of table build, must be a new name

description

Description of the new table

color_variable

Annotation column to use for coloring plotted motifs, provided by counts table from input result's musica object

color_mapping

Mapping from color_variable to color names, provided by counts table from input result's musica object

overwrite

Overwrite existing count table

Value

None

Examples

g <- select_genome("19")

data(musica)
build_standard_table(musica, g, "SBS96", overwrite = TRUE)
#> Building count table from SBS with SBS96 schema
#> Warning: Overwriting counts table: SBS96

annotate_transcript_strand(musica, "19")
#>   403 genes were dropped because they have exons located on both strands
#>   of the same reference sequence or on more than one reference sequence,
#>   so cannot be represented by a single genomic range.
#>   Use 'single.strand.genes.only=FALSE' to get all the genes in a
#>   GRangesList object, or use suppressMessages() to suppress this message.
build_standard_table(musica, g, "SBS192", "Transcript_Strand")
#> Building count table from SBS and Transcript_Strand with SBS192 schema

combine_count_tables(musica, c("SBS96", "SBS192_Trans"), "combo")