Input a cancer subtype to return a list of related COSMIC signatures

Input a cancer subtype to return a list of related COSMIC signatures

Usage

cosmic_v2_subtype_map(tumor_type)

Arguments

tumor_type

Cancer subtype to view related signatures

Value

Returns signatures related to a partial string match

Examples

cosmic_v2_subtype_map("lung")
#> lung adeno
#> 124561317
#> lung  small cell
#> 14515
#> lung  squamous
#> 124513