Create a UMAP from a model result

Proportional sample exposures will be used as input into the umap function to generate a two dimensional UMAP.

Usage

create_umap(
  musica,
  model_name,
  modality = "SBS96",
  result_name = "result",
  n_neighbors = 30,
  min_dist = 0.75,
  spread = 1
)

Arguments

musica

A musica object containing a mutational signature discovery or prediction.

model_name

The name of the desired model.

modality

The modality of the model. Must be "SBS96", "DBS78", or "IND83". Default "SBS96".

result_name

Name of the result list entry containing the model. Default "result".

n_neighbors

The size of local neighborhood used for views of manifold approximation. Larger values result in more global the manifold, while smaller values result in more local data being preserved. If n_neighbors is larger than the number of samples, then n_neighbors will automatically be set to the number of samples in the musica. Default 30.

min_dist

The effective minimum distance between embedded points. Smaller values will result in a more clustered/clumped embedding where nearby points on the manifold are drawn closer together, while larger values will result on a more even dispersal of points. Default 0.2.

spread

The effective scale of embedded points. In combination with ‘min_dist’, this determines how clustered/clumped the embedded points are. Default 1.

Value

A musica object with a new UMAP stored in the UMAP slot of the result_model object for the model.

See also

Examples

data(res_annot)
create_umap(res_annot, model_name = "res_annot")
#> The parameter 'n_neighbors' cannot be bigger than the total number of samples. Setting 'n_neighbors' to 7.