Compare exposures of annotated samples

exposure_differential_analysis is used to run differential analysis on the signature exposures of annotated samples.

Usage

exposure_differential_analysis(
  musica,
  model_name,
  annotation,
  modality = "SBS96",
  result_name = "result",
  method = c("wilcox", "kruskal", "glm.nb"),
  group1 = NULL,
  group2 = NULL,
  ...
)

Arguments

musica

A musica object.

model_name

The name of the model.

annotation

Column in the sample_annotations table of the musica object

modality

The modality. Must be "SBS96", "DBS78", or "IND83". Default "SBS96".

result_name

Name of the result list entry. Default "result".

method

Any method in c("wilcox", "kruskal", "glm.nb") used to perform differential analysis on signature exposures

group1

character vector used in the Wilcox test. Elements in group1 are compared to elements in group2. This is required for annotation with more than 2 levels.

group2

character vector used in the Wilcox test. Elements in group2 are compared to elements in group1. This is required for annotation with more than 2 levels.

...

Additional arguments to be passed to the chosen method

Value

A matrix containing statistics summarizing the analysis dependent on the chosen method

Examples

data("res_annot")
exposure_differential_analysis(res_annot,
  model_name = "res_annot",
  annotation = "Tumor_Subtypes", method = "wilcox"
)
#>    Signature Group1 Group2          Pr(>|z|)             (fdr)
#> 1 Signature1 Breast   Lung 0.114285714285714 0.342857142857143
#> 2 Signature2 Breast   Lung 0.228571428571429 0.342857142857143
#> 3 Signature3 Breast   Lung 0.628571428571429 0.628571428571429