Generate result_grid from musica based on annotation and range of k

Generate result_grid from musica based on annotation and range of k

Usage

generate_result_grid(
  musica,
  modality,
  algorithm = "lda",
  annotation = NA,
  k_start,
  k_end,
  result_name = "result_grid",
  n_start = 1,
  seed = NULL,
  par_cores = FALSE,
  verbose = FALSE,
  make_copy = FALSE,
  table_name = NULL
)

Arguments

musica

A musica object.

modality

Modality used for signature discovery

algorithm

Algorithm for signature discovery

annotation

Sample annotation to split results into

k_start

Lower range of number of signatures for discovery

k_end

Upper range of number of signatures for discovery

result_name

Name for result_list entry to save the results to. Default "result_grid".

n_start

Number of times to discover signatures and compare based on posterior loglikihood

seed

Seed to use for reproducible results, set to null to disable

par_cores

Number of parallel cores to use (NMF only)

verbose

Whether to output loop iterations

make_copy

If FALSE, the inputted musica object is updated and nothing is returned. If TRUE, a new musica object is created and returned. Default FALSE.

table_name

Use modality instead

Value

Returns nothing or a new musica object, depending on the make_copy parameter.

Examples

data(musica_sbs96)
grid <- generate_result_grid(musica_sbs96, "SBS96", "lda",
  k_start = 2,
  k_end = 5
)