Object containing deconvolved/predicted signatures, sample weights, and the musica object the result was generated from

Object containing deconvolved/predicted signatures, sample weights, and the musica object the result was generated from

Slots

signatures

A matrix of signatures by mutational motifs

exposures

A matrix of samples by signature weights

table_name

A character vector of table names used to make the result

algorithm

Describes how the signatures/weights were generated

musica

The musica object the results were generated from

umap

List of umap data.frames for plotting and analysis