Retrieve signatures from a result_model, result_collection, or musica object

The signatures matrix contains the probability of mutation motif in each sample. Rows correspond to each motif and columns correspond to each signature.

Usage

signatures(x, ...)

# S4 method for class 'musica'
signatures(x, result, modality, model_id)

# S4 method for class 'result_collection'
signatures(x, modality, model_id)

# S4 method for class 'result_model'
signatures(x)

signatures(x, ...) <- value

# S4 method for class 'musica,matrix'
signatures(x, result, modality, model_id) <- value

# S4 method for class 'result_collection,matrix'
signatures(x, modality, model_id) <- value

# S4 method for class 'result_model,matrix'
signatures(x) <- value

Arguments

x

A musica, result_collection, or result_model object generated by a mutational discovery or prediction tool.

...

Other inputs

result

Name of result from result_list to assign the signatures. Used when result is a musica object.

modality

Modality to assign the signatures. Used when result is a musica or result_collection object.

model_id

Model identifier to assign the signatures. Used when result is a musica or result_collection object.

value

A matrix of motifs counts by samples

Value

A matrix of mutational signatures

Examples

data(res)
signatures(res, "result", "SBS96", "res")
#>           Signature1   Signature2   Signature3
#> C>A_ACA 8.666532e-04 2.613711e-02 4.448169e-14
#> C>A_ACC 1.222204e-02 1.367763e-02 3.861026e-17
#> C>A_ACG 8.228520e-03 7.525241e-07 1.181972e-34
#> C>A_ACT 6.922897e-03 1.046267e-02 4.705594e-16
#> C>A_CCA 2.268506e-02 2.627552e-02 5.342028e-03
#> C>A_CCC 3.453916e-02 1.326883e-02 8.573563e-19
#> C>A_CCG 3.017375e-02 4.968338e-12 4.942740e-34
#> C>A_CCT 2.347529e-02 7.349768e-03 7.630364e-34
#> C>A_GCA 1.767337e-02 3.157788e-02 5.691025e-03
#> C>A_GCC 1.424257e-02 1.748020e-02 1.014869e-34
#> C>A_GCG 1.285337e-64 1.203912e-02 9.483122e-67
#> C>A_GCT 8.555769e-03 1.168081e-02 1.683258e-17
#> C>A_TCA 2.488531e-02 1.182221e-02 3.015327e-13
#> C>A_TCC 2.466496e-02 1.507377e-02 4.714786e-34
#> C>A_TCG 1.126220e-02 8.711230e-03 1.727722e-33
#> C>A_TCT 2.454132e-02 1.520943e-02 4.301613e-17
#> C>G_ACA 3.720076e-44 3.720076e-44 3.720076e-44
#> C>G_ACC 2.595663e-05 1.502043e-02 5.342027e-03
#> C>G_ACG 2.038874e-03 6.792227e-03 1.193386e-33
#> C>G_ACT 2.742966e-03 1.119504e-07 1.264509e-34
#> C>G_CCA 9.639451e-03 1.049176e-02 1.171358e-33
#> C>G_CCC 1.371533e-02 1.498665e-08 2.995019e-16
#> C>G_CCG 4.105876e-64 6.019562e-03 5.360183e-66
#> C>G_CCT 4.107476e-03 7.532273e-03 1.634147e-34
#> C>G_GCA 1.891476e-03 1.297352e-02 1.814603e-33
#> C>G_GCC 3.726155e-04 1.163028e-02 3.489254e-35
#> C>G_GCG 8.229101e-03 1.150110e-07 3.013128e-13
#> C>G_GCT 1.097217e-02 1.085797e-07 1.720778e-13
#> C>G_TCA 7.467141e-02 2.040074e-02 1.606676e-13
#> C>G_TCC 3.057653e-02 5.577622e-03 2.153462e-12
#> C>G_TCG 1.097205e-02 2.506732e-07 5.295908e-34
#> C>G_TCT 7.448028e-02 1.760067e-02 8.285057e-13
#> C>T_ACA 6.403196e-03 2.307202e-02 1.793321e-35
#> C>T_ACC 3.020069e-03 1.670874e-02 1.254062e-02
#> C>T_ACG 1.491455e-02 3.459969e-02 5.700336e-03
#> C>T_ACT 6.830307e-03 7.554477e-03 4.916733e-15
#> C>T_CCA 2.419367e-02 3.456344e-02 6.248236e-02
#> C>T_CCC 2.118356e-02 1.286786e-02 1.335619e-01
#> C>T_CCG 2.102095e-02 5.086909e-02 2.903630e-02
#> C>T_CCT 6.045208e-03 3.811245e-02 8.084269e-02
#> C>T_GCA 1.574014e-03 3.439032e-02 3.551824e-15
#> C>T_GCC 1.777753e-02 2.723358e-02 2.922491e-02
#> C>T_GCG 2.774451e-02 2.034235e-02 6.019646e-03
#> C>T_GCT 4.098337e-04 3.431291e-02 7.746617e-03
#> C>T_TCA 3.189400e-02 3.519934e-02 9.443587e-02
#> C>T_TCC 2.449972e-02 2.107020e-02 2.354123e-01
#> C>T_TCG 2.219426e-02 3.989305e-02 1.049950e-01
#> C>T_TCT 4.067045e-02 2.032979e-02 3.963147e-02
#> T>A_ATA 2.373571e-03 4.050712e-04 6.268995e-07
#> T>A_ATC 4.600210e-03 3.981849e-03 5.342031e-03
#> T>A_ATG 5.030936e-03 1.253859e-02 4.563086e-26
#> T>A_ATT 5.486136e-03 3.009781e-03 5.342034e-03
#> T>A_CTA 9.013459e-03 5.159054e-03 4.092394e-33
#> T>A_CTC 1.097227e-02 1.058355e-09 3.709747e-10
#> T>A_CTG 2.386812e-02 9.928521e-03 5.342028e-03
#> T>A_CTT 5.172173e-04 1.749118e-02 8.753248e-34
#> T>A_GTA 1.483420e-73 6.598148e-70 1.068406e-02
#> T>A_GTC 2.743066e-03 3.026769e-09 3.172478e-35
#> T>A_GTG 5.485563e-03 6.290277e-07 5.342033e-03
#> T>A_GTT 8.953417e-04 1.406651e-02 7.645306e-15
#> T>A_TTA 7.830455e-03 1.726050e-08 6.118554e-03
#> T>A_TTC 3.825787e-65 3.009781e-03 4.022010e-67
#> T>A_TTG 8.229205e-03 2.408105e-13 4.092867e-33
#> T>A_TTT 1.097650e-04 5.899124e-03 5.342033e-03
#> T>C_ATA 1.369925e-02 1.766050e-05 5.069676e-10
#> T>C_ATC 5.486137e-03 3.762914e-62 5.342032e-03
#> T>C_ATG 8.925114e-03 2.632446e-02 7.800215e-16
#> T>C_ATT 2.433390e-03 9.728480e-03 1.538829e-02
#> T>C_CTA 4.600669e-03 3.981344e-03 2.719501e-34
#> T>C_CTC 2.561366e-06 9.026533e-03 2.172342e-35
#> T>C_CTG 1.378588e-02 1.196173e-02 2.980716e-14
#> T>C_CTT 2.722357e-65 6.019562e-03 4.772865e-67
#> T>C_GTA 1.096049e-02 1.292396e-05 8.337786e-10
#> T>C_GTC 1.097207e-02 2.201382e-07 5.033614e-10
#> T>C_GTG 7.807325e-03 2.401830e-05 6.121000e-03
#> T>C_GTT 4.598791e-03 3.983399e-03 2.136814e-02
#> T>C_TTA 1.877293e-05 1.152941e-02 6.210160e-03
#> T>C_TTC 2.666685e-04 1.744657e-02 5.909145e-03
#> T>C_TTG 3.362178e-03 7.928723e-03 6.089818e-03
#> T>C_TTT 7.719903e-05 5.934856e-03 1.068407e-02
#> T>G_ATA 5.485846e-03 3.193990e-07 4.766222e-34
#> T>G_ATC 2.742837e-03 2.539378e-07 5.342032e-03
#> T>G_ATG 2.209459e-71 1.284586e-67 5.342032e-03
#> T>G_ATT 6.715116e-03 4.671088e-03 1.217154e-33
#> T>G_CTA 2.206821e-79 7.231800e-78 5.342032e-03
#> T>G_CTC 6.752989e-03 7.639313e-03 8.133268e-15
#> T>G_CTG 2.742934e-03 1.470164e-07 5.342033e-03
#> T>G_CTT 5.485469e-03 7.326000e-07 3.062070e-34
#> T>G_GTA 1.855451e-04 5.815976e-03 2.779645e-34
#> T>G_GTC 3.720076e-44 3.720076e-44 3.720076e-44
#> T>G_GTG 5.486137e-03 1.452622e-12 6.501419e-35
#> T>G_GTT 2.497966e-65 3.009781e-03 1.177083e-67
#> T>G_TTA 1.975414e-65 3.009781e-03 2.176097e-67
#> T>G_TTC 3.720076e-44 3.720076e-44 3.720076e-44
#> T>G_TTG 8.141805e-03 9.589828e-05 1.232222e-12
#> T>G_TTT 2.359148e-03 4.208018e-04 7.955750e-07
data(res)
signatures(res, "result", "SBS96", "res") <- matrix()