Retrieve the list of count_tables from a musica or musica_result object

The count_tables contains standard and/or custom count tables created from variants

tables(object)

# S4 method for musica
tables(object)

# S4 method for musica_result
tables(object)

tables(musica) <- value

# S4 method for musica,list
tables(musica) <- value

Arguments

object
A musica object generated by the create_musica function or a musica_result object generated by a mutational discovery or prediction tool.

musica
A musica object generated by the create_musica function or a musica_result object generated by a mutational discovery or prediction tool.

value
A list of count_table objects representing counts of motifs in samples

Value

A list of count_tables

Examples

data(res)
tables(res)
#> $SBS96
#> Count_Table:  SBS96 
#> Motifs: 96 
#> Samples: 7 
#>  
#> **Annotations: 
#>            motif mutation context
#> C>A_ACA C>A_ACA      C>A     ACA
#> C>A_ACC C>A_ACC      C>A     ACC
#> C>A_ACG C>A_ACG      C>A     ACG
#> C>A_ACT C>A_ACT      C>A     ACT
#> C>A_CCA C>A_CCA      C>A     CCA
#> C>A_CCC C>A_CCC      C>A     CCC
#> 7           ...      ...     ... 
#> 
#> **Features: 
#>    mutation
#> 1  C>A_TCA
#> 2  C>A_ACT
#> 3  T>C_ATT
#> 4  C>G_TCT
#> 5  T>G_CTC
#> 6  C>G_CCC
#> 7      ... 
#> 
#> **Types: 
#>  SBS
#>  
#> **Color Variable: 
#>  mutation
#>  
#> **Color Mapping: 
#>  #5ABCEBFF
#>  #050708FF
#>  #D33C32FF
#>  #CBCACBFF
#>  #ABCD72FF
#>  #E7C9C6FF
#>  
#> **Descriptions: 
#>  Single Base Substitution table with one base upstream and downstream
#> 
data(musica)
tables(musica)
#> $SBS96
#> Count_Table:  SBS96 
#> Motifs: 96 
#> Samples: 7 
#>  
#> **Annotations: 
#>            motif mutation context
#> C>A_ACA C>A_ACA      C>A     ACA
#> C>A_ACC C>A_ACC      C>A     ACC
#> C>A_ACG C>A_ACG      C>A     ACG
#> C>A_ACT C>A_ACT      C>A     ACT
#> C>A_CCA C>A_CCA      C>A     CCA
#> C>A_CCC C>A_CCC      C>A     CCC
#> 7           ...      ...     ... 
#> 
#> **Features: 
#>    mutation
#> 1  C>A_TCA
#> 2  C>A_ACT
#> 3  T>C_ATT
#> 4  C>G_TCT
#> 5  T>G_CTC
#> 6  C>G_CCC
#> 7      ... 
#> 
#> **Types: 
#>  SBS
#>  
#> **Color Variable: 
#>  mutation
#>  
#> **Color Mapping: 
#>  #5ABCEBFF
#>  #050708FF
#>  #D33C32FF
#>  #CBCACBFF
#>  #ABCD72FF
#>  #E7C9C6FF
#>  
#> **Descriptions: 
#>  Single Base Substitution table with one base upstream and downstream
#>