Automatic filtering of signatures for exposure prediction gridded across specific annotation

Automatic filtering of signatures for exposure prediction gridded across specific annotation

auto_predict_grid(
  musica,
  table_name,
  signature_res,
  algorithm,
  sample_annotation = NULL,
  min_exists = 0.05,
  proportion_samples = 0.25,
  rare_exposure = 0.4,
  verbose = TRUE,
  combine_res = TRUE
)

Arguments

musica
Input samples to predit signature weights

table_name
Name of table used for posterior prediction (e.g. SBS96)

signature_res
Signatures to automatically subset from for prediction

algorithm
Algorithm to use for prediction. Choose from "lda_posterior", decompTumor2Sig, and deconstructSigs

sample_annotation
Annotation to grid across, if none given, prediction subsetting on all samples together

min_exists
Threshold to consider a signature active in a sample

proportion_samples
Threshold of samples to consider a signature active in the cohort

rare_exposure
A sample will be considered active in the cohort if at least one sample has more than this threshold proportion

verbose
Print current annotation value being predicted on

combine_res
Automatically combines a list of annotation results into a single result object with zero exposure values for signatures not found in a given annotation's set of samples

Value

A list of results, one per unique annotation value, if no annotation value is given, returns a single result for all samples, or combines into a single result if combines_res = TRUE

Examples

data(musica_annot)
data(cosmic_v2_sigs)
auto_predict_grid(musica = musica_annot, table_name = "SBS96",
signature_res = cosmic_v2_sigs, algorithm = "lda",
sample_annotation = "Tumor_Subtypes")

#> Lung

#> Breast

#> An object of class "musica_result"
#> Slot "signatures":
#>           Signature1  Signature13  Signature18  Signature20 Signature21
#> C>A_ACA 1.109833e-02 3.347572e-04 5.053642e-02 1.179962e-03      0.0001
#> C>A_ACC 9.149341e-03 6.487361e-04 1.093982e-02 2.211505e-03      0.0007
#> C>A_ACG 1.490070e-03 3.814459e-05 2.288073e-03 1.616910e-07      0.0000
#> C>A_ACT 6.233885e-03 8.466585e-04 1.942409e-02 3.008010e-03      0.0006
#> C>A_CCA 6.595870e-03 1.710090e-03 8.876811e-02 1.737711e-02      0.0020
#> C>A_CCC 7.342368e-03 1.159257e-03 2.064139e-02 3.650246e-02      0.0014
#> C>A_CCG 8.928404e-04 2.441665e-04 1.717840e-02 1.248259e-02      0.0027
#> C>A_CCT 7.186582e-03 1.256768e-03 3.767696e-02 1.034012e-01      0.0056
#> C>A_GCA 8.232604e-03 1.321096e-04 1.287242e-01 1.116124e-03      0.0001
#> C>A_GCC 5.758021e-03 7.542440e-04 1.609288e-02 3.271429e-03      0.0038
#> C>A_GCG 6.163352e-04 9.682078e-05 9.271417e-03 3.751613e-03      0.0003
#> C>A_GCT 4.459080e-03 5.511037e-04 7.201707e-02 1.293719e-02      0.0023
#> C>A_TCA 1.225006e-02 4.816605e-02 6.708064e-02 6.882997e-04      0.0001
#> C>A_TCC 1.116223e-02 1.732956e-02 4.186901e-02 4.058229e-03      0.0005
#> C>A_TCG 2.275496e-03 2.293157e-03 1.516238e-02 1.001102e-03      0.0000
#> C>A_TCT 1.525910e-02 1.863458e-02 1.178778e-01 1.222910e-02      0.0004
#> C>G_ACA 1.801068e-03 3.775165e-03 1.516888e-03 6.970411e-04      0.0005
#> C>G_ACC 2.580909e-03 9.208248e-04 2.498785e-03 2.059310e-03      0.0008
#> C>G_ACG 5.925480e-04 1.989049e-05 2.614509e-03 1.273496e-05      0.0000
#> C>G_ACT 2.963986e-03 3.860632e-03 3.983011e-03 8.485871e-04      0.0018
#> C>G_CCA 1.284983e-03 5.333651e-03 4.584638e-04 1.331626e-05      0.0001
#> C>G_CCC 7.021348e-04 9.123035e-04 3.654228e-03 5.481447e-04      0.0000
#> C>G_CCG 5.062896e-04 2.854673e-04 5.405774e-03 0.000000e+00      0.0000
#> C>G_CCT 1.381543e-03 6.383098e-03 5.519605e-03 1.207306e-04      0.0003
#> C>G_GCA 6.021227e-04 1.349420e-03 9.238797e-04 1.339769e-03      0.0004
#> C>G_GCC 2.393352e-03 7.888077e-04 6.486479e-04 1.811151e-03      0.0042
#> C>G_GCG 2.485340e-07 8.179121e-05 5.428400e-04 6.375380e-07      0.0000
#> C>G_GCT 8.900807e-04 2.680123e-03 1.376725e-03 1.238264e-03      0.0004
#> C>G_TCA 1.874853e-03 2.802349e-01 1.802969e-03 2.429847e-05      0.0002
#> C>G_TCC 2.067419e-03 6.388529e-02 7.863007e-03 1.519986e-04      0.0003
#> C>G_TCG 3.048970e-04 9.528486e-03 1.513223e-03 0.000000e+00      0.0000
#> C>G_TCT 3.151574e-03 3.302255e-01 3.567102e-03 6.830600e-04      0.0003
#> C>T_ACA 2.951453e-02 1.480181e-03 3.825624e-03 3.285805e-02      0.0051
#> C>T_ACC 1.432275e-02 4.050871e-05 2.465214e-03 2.219605e-02      0.0062
#> C>T_ACG 1.716469e-01 3.165947e-04 8.463957e-03 3.457201e-02      0.0256
#> C>T_ACT 1.262376e-02 0.000000e+00 5.605740e-03 1.881956e-02      0.0030
#> C>T_CCA 2.089645e-02 3.647278e-03 9.767511e-03 1.114519e-02      0.0034
#> C>T_CCC 1.850170e-02 1.004229e-03 5.654828e-03 1.815275e-02      0.0091
#> C>T_CCG 9.557722e-02 1.292021e-04 1.618322e-02 4.438494e-02      0.0318
#> C>T_CCT 1.711331e-02 6.117700e-04 5.153319e-03 1.384714e-02      0.0040
#> C>T_GCA 2.494381e-02 2.934138e-03 2.703069e-03 5.679395e-02      0.0163
#> C>T_GCC 2.716149e-02 2.515171e-03 4.199308e-03 5.442986e-02      0.0392
#> C>T_GCG 1.035708e-01 3.205265e-03 1.189736e-02 4.592689e-03      0.0019
#> C>T_GCT 1.768985e-02 1.687731e-03 1.913381e-03 4.958086e-02      0.0166
#> C>T_TCA 1.449210e-02 1.138421e-01 1.492339e-02 1.071032e-02      0.0064
#> C>T_TCC 1.768078e-02 1.502484e-02 9.962744e-03 8.175510e-03      0.0086
#> C>T_TCG 7.600222e-02 6.102098e-03 1.154972e-02 2.316390e-02      0.0126
#> C>T_TCT 1.376170e-02 2.814066e-02 5.398014e-03 9.540087e-03      0.0064
#> T>A_ATA 4.021520e-03 1.243618e-03 3.037211e-03 5.744014e-04      0.0083
#> T>A_ATC 2.371144e-03 1.060002e-03 8.111672e-04 1.971462e-03      0.0292
#> T>A_ATG 2.810910e-03 1.025544e-04 4.503207e-03 1.905866e-04      0.0027
#> T>A_ATT 8.360909e-03 9.718121e-04 2.736523e-03 7.421104e-03      0.0043
#> T>A_CTA 1.182587e-03 2.851590e-04 4.123464e-04 4.504236e-05      0.0000
#> T>A_CTC 1.903167e-03 2.222873e-04 2.491411e-03 4.389056e-03      0.0037
#> T>A_CTG 1.487961e-03 2.869518e-04 5.523046e-03 1.127836e-03      0.0025
#> T>A_CTT 2.179344e-03 8.004460e-07 3.313713e-03 2.385541e-03      0.0010
#> T>A_GTA 6.892894e-04 1.379604e-04 1.875434e-03 1.631298e-04      0.0000
#> T>A_GTC 5.524095e-04 2.528674e-04 1.510310e-03 3.799180e-03      0.0142
#> T>A_GTG 1.200229e-03 1.535175e-04 1.505369e-03 8.283337e-04      0.0031
#> T>A_GTT 2.107137e-03 1.227456e-04 3.611025e-03 1.900747e-03      0.0041
#> T>A_TTA 5.600155e-03 4.032613e-04 4.663278e-03 9.245480e-04      0.0027
#> T>A_TTC 1.999079e-03 7.718133e-04 2.578156e-03 7.082068e-04      0.0045
#> T>A_TTG 1.090066e-03 2.451942e-04 2.016738e-03 9.700313e-05      0.0000
#> T>A_TTT 3.981023e-03 1.603460e-04 2.818679e-03 2.313489e-03      0.0008
#> T>C_ATA 1.391577e-02 5.494542e-04 2.933054e-03 1.639506e-02      0.0265
#> T>C_ATC 6.274961e-03 1.455067e-04 1.513285e-03 1.925179e-02      0.0165
#> T>C_ATG 1.013764e-02 2.791780e-04 3.329669e-03 5.478492e-02      0.0224
#> T>C_ATT 9.256316e-03 2.674614e-04 5.491318e-03 7.341270e-03      0.0189
#> T>C_CTA 4.176675e-03 4.275147e-05 1.921693e-04 1.486770e-02      0.0219
#> T>C_CTC 5.252593e-03 4.330754e-05 3.050193e-03 1.835616e-02      0.0100
#> T>C_CTG 7.013225e-03 5.427253e-05 2.596464e-03 7.415056e-02      0.0211
#> T>C_CTT 6.713813e-03 7.639440e-07 1.310029e-03 1.100112e-02      0.0204
#> T>C_GTA 1.124784e-02 1.054081e-03 5.291265e-04 2.469317e-03      0.1426
#> T>C_GTC 6.999724e-03 5.298997e-04 2.279131e-03 1.565633e-03      0.0665
#> T>C_GTG 4.977593e-03 6.300957e-04 5.871057e-03 2.734176e-02      0.0884
#> T>C_GTT 1.066741e-02 6.761879e-04 4.771566e-03 1.498265e-03      0.1003
#> T>C_TTA 8.073616e-03 7.840561e-04 3.023036e-04 5.734230e-03      0.0389
#> T>C_TTC 4.857381e-03 1.329749e-04 6.260791e-03 1.123425e-02      0.0313
#> T>C_TTG 8.325454e-03 3.740240e-05 1.541265e-03 2.376761e-02      0.0249
#> T>C_TTT 6.257106e-03 1.797028e-04 1.357797e-03 5.395186e-03      0.0274
#> T>G_ATA 1.587636e-03 3.861998e-04 3.861382e-04 1.765291e-06      0.0000
#> T>G_ATC 1.784091e-03 1.123832e-04 1.254982e-03 3.813879e-04      0.0031
#> T>G_ATG 1.385831e-03 3.600070e-05 2.369216e-03 6.587581e-04      0.0016
#> T>G_ATT 3.158539e-03 1.924938e-04 3.386029e-04 5.321252e-04      0.0018
#> T>G_CTA 3.026912e-04 1.509421e-05 6.100255e-05 2.326408e-05      0.0000
#> T>G_CTC 2.098502e-03 1.476248e-04 2.015697e-03 8.476495e-03      0.0005
#> T>G_CTG 1.599549e-03 4.598730e-05 5.311967e-03 1.538378e-02      0.0004
#> T>G_CTT 2.758538e-03 4.637147e-04 2.341629e-03 2.185395e-03      0.0032
#> T>G_GTA 9.904500e-05 1.848986e-05 1.341532e-03 3.579149e-06      0.0000
#> T>G_GTC 2.023656e-04 9.337351e-05 9.204310e-07 2.830575e-03      0.0006
#> T>G_GTG 1.188353e-03 1.057919e-05 9.695000e-03 2.733219e-03      0.0004
#> T>G_GTT 8.007233e-04 5.374121e-05 2.890554e-03 3.558559e-03      0.0008
#> T>G_TTA 1.397554e-03 5.465818e-04 3.631840e-05 2.691470e-07      0.0000
#> T>G_TTC 1.291737e-03 2.353471e-04 3.233838e-03 3.772344e-04      0.0018
#> T>G_TTG 2.031077e-03 4.790220e-07 7.546018e-04 5.154216e-04      0.0003
#> T>G_TTT 4.030128e-03 6.705883e-04 2.126442e-03 6.156567e-04      0.0003
#>          Signature22  Signature24   Signature3 Signature30 Signature4
#> C>A_ACA 0.0015040704 2.864599e-02 0.0221723068   0.0000000     0.0365
#> C>A_ACC 0.0024510110 2.021464e-02 0.0178716754   0.0000000     0.0309
#> C>A_ACG 0.0000000000 2.047900e-02 0.0021383396   0.0019673     0.0183
#> C>A_ACT 0.0009224525 2.460015e-02 0.0162651456   0.0000000     0.0243
#> C>A_CCA 0.0045496929 6.355928e-02 0.0187817256   0.0000000     0.0461
#> C>A_CCC 0.0037644739 3.375700e-02 0.0157604578   0.0000000     0.0614
#> C>A_CCG 0.0009001633 2.242899e-02 0.0019633898   0.0022624     0.0088
#> C>A_CCT 0.0044398462 2.008652e-02 0.0147228611   0.0000000     0.0432
#> C>A_GCA 0.0012983702 5.467645e-02 0.0096965397   0.0088530     0.0376
#> C>A_GCC 0.0018096222 1.099460e-01 0.0108433411   0.0093449     0.0399
#> C>A_GCG 0.0003338381 4.652487e-02 0.0009291405   0.0008853     0.0227
#> C>A_GCT 0.0005191482 5.042871e-02 0.0122153826   0.0081645     0.0258
#> C>A_TCA 0.0040516346 3.145126e-02 0.0116532249   0.0000000     0.0330
#> C>A_TCC 0.0027397709 6.253829e-02 0.0166067750   0.0000000     0.0538
#> C>A_TCG 0.0005131511 1.280259e-02 0.0013572394   0.0016722     0.0104
#> C>A_TCT 0.0036056509 3.029768e-02 0.0163280760   0.0000000     0.0370
#> C>G_ACA 0.0005276440 1.199304e-02 0.0240026152   0.0000000     0.0097
#> C>G_ACC 0.0000000000 8.425120e-03 0.0121603037   0.0000000     0.0054
#> C>G_ACG 0.0000000000 0.000000e+00 0.0052754195   0.0048200     0.0031
#> C>G_ACT 0.0002994547 3.881383e-03 0.0232776563   0.0000000     0.0054
#> C>G_CCA 0.0005820177 1.347682e-02 0.0168325720   0.0000000     0.0105
#> C>G_CCC 0.0008252996 1.610569e-02 0.0135314415   0.0000000     0.0097
#> C>G_CCG 0.0003821146 7.974720e-04 0.0041764575   0.0032461     0.0063
#> C>G_CCT 0.0017606459 1.038867e-02 0.0240463904   0.0000000     0.0094
#> C>G_GCA 0.0000000000 4.945061e-03 0.0119168446   0.0073775     0.0070
#> C>G_GCC 0.0000000000 9.369876e-03 0.0098236534   0.0065906     0.0091
#> C>G_GCG 0.0002146959 5.625727e-03 0.0016710543   0.0015739     0.0062
#> C>G_GCT 0.0005384388 3.055770e-03 0.0179143343   0.0000000     0.0060
#> C>G_TCA 0.0002254906 1.022086e-02 0.0160414054   0.0000000     0.0032
#> C>G_TCC 0.0006286950 7.578365e-03 0.0201499204   0.0000000     0.0105
#> C>G_TCG 0.0003958080 1.592714e-03 0.0025279109   0.0019673     0.0031
#> C>G_TCT 0.0002573751 6.100727e-03 0.0326736408   0.0000000     0.0050
#> C>T_ACA 0.0031133897 6.315518e-03 0.0178721707   0.0651190     0.0120
#> C>T_ACC 0.0015248603 3.844327e-03 0.0088960343   0.0543970     0.0075
#> C>T_ACG 0.0036256412 3.110664e-03 0.0035727083   0.0204604     0.0028
#> C>T_ACT 0.0000000000 3.420419e-03 0.0147976122   0.0219359     0.0059
#> C>T_CCA 0.0077313495 6.035447e-03 0.0143950595   0.0694472     0.0210
#> C>T_CCC 0.0071108505 0.000000e+00 0.0085447812   0.0638403     0.0144
#> C>T_CCG 0.0055020310 2.924419e-03 0.0035184658   0.0173126     0.0076
#> C>T_CCT 0.0041697772 3.971202e-03 0.0160755457   0.0342318     0.0201
#> C>T_GCA 0.0019153709 2.367905e-02 0.0161272733   0.0485934     0.0087
#> C>T_GCC 0.0032291335 1.591920e-02 0.0082086324   0.0494787     0.0080
#> C>T_GCG 0.0033005989 1.264808e-02 0.0012129229   0.0157387     0.0023
#> C>T_GCT 0.0011140596 1.812978e-02 0.0106116492   0.0188865     0.0082
#> C>T_TCA 0.0020167217 2.657189e-02 0.0088802091   0.0849892     0.0035
#> C>T_TCC 0.0019908342 1.205569e-02 0.0135295728   0.0903010     0.0070
#> C>T_TCG 0.0014984731 5.363360e-03 0.0017054089   0.0151485     0.0011
#> C>T_TCT 0.0011822265 5.532499e-03 0.0103044169   0.0455440     0.0077
#> T>A_ATA 0.0496192297 1.166747e-03 0.0084285636   0.0075743     0.0048
#> T>A_ATC 0.0116667405 1.517306e-05 0.0073730967   0.0037380     0.0039
#> T>A_ATG 0.0679154542 6.332456e-03 0.0073573000   0.0065906     0.0100
#> T>A_ATT 0.0145999176 4.967216e-06 0.0087539390   0.0093447     0.0030
#> T>A_CTA 0.0868985455 2.766635e-03 0.0075713203   0.0053118     0.0075
#> T>A_CTC 0.0537068468 2.694462e-03 0.0127254635   0.0067873     0.0111
#> T>A_CTG 0.2132208666 6.911165e-03 0.0115085619   0.0082628     0.0342
#> T>A_CTT 0.0618275075 2.915200e-03 0.0164561777   0.0000000     0.0115
#> T>A_GTA 0.0537091457 2.225127e-03 0.0044347826   0.0037380     0.0069
#> T>A_GTC 0.0094893967 4.853149e-03 0.0056154324   0.0033445     0.0052
#> T>A_GTG 0.0502939024 1.957586e-04 0.0080701365   0.0050167     0.0133
#> T>A_GTT 0.0106014572 0.000000e+00 0.0086791217   0.0063939     0.0045
#> T>A_TTA 0.0408105028 0.000000e+00 0.0071326814   0.0076726     0.0045
#> T>A_TTC 0.0255703958 7.994129e-03 0.0091027978   0.0064922     0.0046
#> T>A_TTG 0.0698452181 1.120214e-02 0.0065658630   0.0036396     0.0082
#> T>A_TTT 0.0235790618 2.705788e-04 0.0147121354   0.0000000     0.0045
#> T>C_ATA 0.0168399309 4.726009e-03 0.0130357094   0.0000000     0.0084
#> T>C_ATC 0.0011082624 6.279493e-04 0.0091862897   0.0087547     0.0020
#> T>C_ATG 0.0049136164 2.723759e-03 0.0117169201   0.0090498     0.0081
#> T>C_ATT 0.0026305239 7.122368e-03 0.0169788708   0.0000000     0.0036
#> T>C_CTA 0.0040273463 1.848877e-03 0.0078950126   0.0000000     0.0052
#> T>C_CTC 0.0013524439 4.453690e-03 0.0144311704   0.0000000     0.0026
#> T>C_CTG 0.0027496661 0.000000e+00 0.0084229721   0.0000000     0.0100
#> T>C_CTT 0.0023633535 1.597389e-03 0.0119324305   0.0000000     0.0054
#> T>C_GTA 0.0042174541 1.269750e-03 0.0068500556   0.0000000     0.0061
#> T>C_GTC 0.0000000000 6.710340e-03 0.0062607490   0.0065906     0.0016
#> T>C_GTG 0.0014748845 2.776850e-04 0.0060987630   0.0078694     0.0042
#> T>C_GTT 0.0015610427 4.238478e-03 0.0075093439   0.0084596     0.0024
#> T>C_TTA 0.0048036697 1.682291e-03 0.0091154951   0.0000000     0.0028
#> T>C_TTC 0.0009121575 2.344528e-03 0.0109537909   0.0091481     0.0016
#> T>C_TTG 0.0017817357 4.812742e-04 0.0061130332   0.0060004     0.0036
#> T>C_TTT 0.0000000000 3.581480e-04 0.0107743363   0.0000000     0.0022
#> T>G_ATA 0.0016519986 0.000000e+00 0.0023513681   0.0053118     0.0000
#> T>G_ATC 0.0000000000 1.463951e-03 0.0014642311   0.0024592     0.0002
#> T>G_ATG 0.0010868728 8.476125e-04 0.0090537776   0.0062955     0.0015
#> T>G_ATT 0.0000000000 1.075455e-03 0.0070309764   0.0095416     0.0002
#> T>G_CTA 0.0011203565 0.000000e+00 0.0019741811   0.0046233     0.0000
#> T>G_CTC 0.0010645836 2.185437e-03 0.0058242955   0.0060004     0.0013
#> T>G_CTG 0.0041885681 1.139097e-04 0.0104646545   0.0073775     0.0046
#> T>G_CTT 0.0016294096 2.533718e-03 0.0087243873   0.0091481     0.0012
#> T>G_GTA 0.0000000000 0.000000e+00 0.0041444880   0.0032461     0.0000
#> T>G_GTC 0.0000000000 1.073632e-03 0.0045019853   0.0018690     0.0000
#> T>G_GTG 0.0018775892 2.924697e-03 0.0163914526   0.0033445     0.0018
#> T>G_GTT 0.0000000000 8.250983e-04 0.0070672366   0.0056069     0.0002
#> T>G_TTA 0.0016516988 0.000000e+00 0.0054271842   0.0086563     0.0000
#> T>G_TTC 0.0000000000 0.000000e+00 0.0061602504   0.0043282     0.0003
#> T>G_TTG 0.0025727520 1.360505e-03 0.0110765263   0.0082628     0.0030
#> T>G_TTT 0.0000000000 6.951578e-05 0.0130009842   0.0000000     0.0011
#>         Signature6 Signature7
#> C>A_ACA     0.0017     0.0004
#> C>A_ACC     0.0028     0.0005
#> C>A_ACG     0.0005     0.0000
#> C>A_ACT     0.0019     0.0004
#> C>A_CCA     0.0101     0.0012
#> C>A_CCC     0.0241     0.0006
#> C>A_CCG     0.0091     0.0000
#> C>A_CCT     0.0571     0.0013
#> C>A_GCA     0.0024     0.0003
#> C>A_GCC     0.0058     0.0001
#> C>A_GCG     0.0021     0.0000
#> C>A_GCT     0.0087     0.0001
#> C>A_TCA     0.0017     0.0010
#> C>A_TCC     0.0029     0.0020
#> C>A_TCG     0.0011     0.0002
#> C>A_TCT     0.0058     0.0013
#> C>G_ACA     0.0013     0.0000
#> C>G_ACC     0.0012     0.0000
#> C>G_ACG     0.0000     0.0000
#> C>G_ACT     0.0018     0.0001
#> C>G_CCA     0.0000     0.0001
#> C>G_CCC     0.0000     0.0004
#> C>G_CCG     0.0000     0.0006
#> C>G_CCT     0.0002     0.0003
#> C>G_GCA     0.0000     0.0000
#> C>G_GCC     0.0030     0.0004
#> C>G_GCG     0.0000     0.0000
#> C>G_GCT     0.0017     0.0000
#> C>G_TCA     0.0000     0.0002
#> C>G_TCC     0.0002     0.0010
#> C>G_TCG     0.0000     0.0003
#> C>G_TCT     0.0001     0.0007
#> C>T_ACA     0.0312     0.0000
#> C>T_ACC     0.0163     0.0197
#> C>T_ACG     0.0908     0.0001
#> C>T_ACT     0.0149     0.0043
#> C>T_CCA     0.0085     0.0754
#> C>T_CCC     0.0099     0.1007
#> C>T_CCG     0.0901     0.0208
#> C>T_CCT     0.0087     0.0788
#> C>T_GCA     0.0653     0.0000
#> C>T_GCC     0.0773     0.0210
#> C>T_GCG     0.1339     0.0002
#> C>T_GCT     0.0524     0.0161
#> C>T_TCA     0.0074     0.1202
#> C>T_TCC     0.0067     0.2887
#> C>T_TCG     0.0391     0.0992
#> C>T_TCT     0.0047     0.0844
#> T>A_ATA     0.0006     0.0010
#> T>A_ATC     0.0033     0.0008
#> T>A_ATG     0.0000     0.0009
#> T>A_ATT     0.0053     0.0035
#> T>A_CTA     0.0001     0.0000
#> T>A_CTC     0.0026     0.0014
#> T>A_CTG     0.0008     0.0007
#> T>A_CTT     0.0011     0.0018
#> T>A_GTA     0.0000     0.0000
#> T>A_GTC     0.0028     0.0000
#> T>A_GTG     0.0006     0.0006
#> T>A_GTT     0.0021     0.0010
#> T>A_TTA     0.0002     0.0010
#> T>A_TTC     0.0008     0.0015
#> T>A_TTG     0.0000     0.0010
#> T>A_TTT     0.0007     0.0050
#> T>C_ATA     0.0075     0.0005
#> T>C_ATC     0.0056     0.0001
#> T>C_ATG     0.0217     0.0007
#> T>C_ATT     0.0023     0.0011
#> T>C_CTA     0.0062     0.0010
#> T>C_CTC     0.0040     0.0008
#> T>C_CTG     0.0270     0.0010
#> T>C_CTT     0.0033     0.0045
#> T>C_GTA     0.0122     0.0002
#> T>C_GTC     0.0059     0.0001
#> T>C_GTG     0.0115     0.0003
#> T>C_GTT     0.0042     0.0081
#> T>C_TTA     0.0059     0.0023
#> T>C_TTC     0.0035     0.0018
#> T>C_TTG     0.0106     0.0019
#> T>C_TTT     0.0029     0.0024
#> T>G_ATA     0.0000     0.0000
#> T>G_ATC     0.0017     0.0000
#> T>G_ATG     0.0007     0.0009
#> T>G_ATT     0.0029     0.0000
#> T>G_CTA     0.0001     0.0000
#> T>G_CTC     0.0040     0.0008
#> T>G_CTG     0.0050     0.0009
#> T>G_CTT     0.0086     0.0013
#> T>G_GTA     0.0000     0.0000
#> T>G_GTC     0.0016     0.0000
#> T>G_GTG     0.0010     0.0017
#> T>G_GTT     0.0035     0.0009
#> T>G_TTA     0.0009     0.0000
#> T>G_TTC     0.0019     0.0010
#> T>G_TTG     0.0011     0.0010
#> T>G_TTT     0.0072     0.0014
#> 
#> Slot "exposures":
#>             TCGA-56-7582-01A-11D-2042-08 TCGA-77-7335-01A-11D-2042-08
#> Signature1                     32.152658                     38.32270
#> Signature13                     8.682772                     50.63988
#> Signature18                    12.082592                      6.29390
#> Signature20                    16.234629                     15.43989
#> Signature21                     0.000000                      0.00000
#> Signature22                     0.000000                      0.00000
#> Signature24                     9.536237                      5.92637
#> Signature3                     22.079051                     34.69719
#> Signature30                    21.289529                     27.48385
#> Signature4                     50.583785                     65.01330
#> Signature6                      8.137026                     12.14373
#> Signature7                     11.221722                     18.03919
#>             TCGA-94-7557-01A-11D-2122-08 TCGA-97-7938-01A-11D-2167-08
#> Signature1                     19.219138                     8.482122
#> Signature13                    30.013888                    13.541006
#> Signature18                     3.325813                    10.951363
#> Signature20                     9.631490                     7.405128
#> Signature21                     0.000000                     0.000000
#> Signature22                     0.000000                     0.000000
#> Signature24                     1.710329                     7.437439
#> Signature3                      9.800487                     8.821790
#> Signature30                     9.006214                     7.117174
#> Signature4                     13.468902                    44.130037
#> Signature6                      6.467568                     2.856502
#> Signature7                     14.356172                     6.257438
#>             TCGA-EE-A3J5-06A-11D-A20D-08 TCGA-ER-A197-06A-32D-A197-08
#> Signature1                      3.731967                     1.579759
#> Signature13                     0.000000                     0.000000
#> Signature18                     0.000000                     0.000000
#> Signature20                     0.000000                     0.000000
#> Signature21                     8.099765                     4.005298
#> Signature22                     5.255852                     1.999809
#> Signature24                     0.000000                     0.000000
#> Signature3                      0.000000                     0.000000
#> Signature30                     0.000000                     0.000000
#> Signature4                      0.000000                     0.000000
#> Signature6                      0.000000                     0.000000
#> Signature7                    104.912416                     3.415135
#>             TCGA-ER-A19O-06A-11D-A197-08
#> Signature1                      5.017192
#> Signature13                     0.000000
#> Signature18                     0.000000
#> Signature20                     0.000000
#> Signature21                     2.344125
#> Signature22                     1.181250
#> Signature24                     0.000000
#> Signature3                      0.000000
#> Signature30                     0.000000
#> Signature4                      0.000000
#> Signature6                      0.000000
#> Signature7                     42.457433
#> 
#> Slot "table_name":
#> [1] "SBS96"
#> 
#> Slot "algorithm":
#> character(0)
#> 
#> Slot "musica":
#> An object of class "musica"
#> Slot "variants":
#>        chr     start       end ref alt                       sample
#>   1:  chr1  11020563  11020563   C   A TCGA-94-7557-01A-11D-2122-08
#>   2:  chr1  43430030  43430030   G   T TCGA-94-7557-01A-11D-2122-08
#>   3:  chr1  58682403  58682403   A   G TCGA-94-7557-01A-11D-2122-08
#>   4:  chr1 109508295 109508295   C   G TCGA-94-7557-01A-11D-2122-08
#>   5:  chr1 156384826 156384826   A   C TCGA-94-7557-01A-11D-2122-08
#>  ---                                                               
#> 907: chr19  54885292  54885292   G   A TCGA-ER-A19O-06A-11D-A197-08
#> 908: chr20  49374328  49374328   A   T TCGA-ER-A19O-06A-11D-A197-08
#> 909:  chrX  73213768  73213768   T   G TCGA-ER-A19O-06A-11D-A197-08
#> 910:  chrX 101292834 101292834   G   A TCGA-ER-A19O-06A-11D-A197-08
#> 911:  chrX 107526690 107526690   C   T TCGA-ER-A19O-06A-11D-A197-08
#>      Variant_Type
#>   1:          SBS
#>   2:          SBS
#>   3:          SBS
#>   4:          SBS
#>   5:          SBS
#>  ---             
#> 907:          SBS
#> 908:          SBS
#> 909:          SBS
#> 910:          SBS
#> 911:          SBS
#> 
#> Slot "count_tables":
#> $SBS96
#> Count_Table:  SBS96 
#> Motifs: 96 
#> Samples: 7 
#>  
#> **Annotations: 
#>            motif mutation context
#> C>A_ACA C>A_ACA      C>A     ACA
#> C>A_ACC C>A_ACC      C>A     ACC
#> C>A_ACG C>A_ACG      C>A     ACG
#> C>A_ACT C>A_ACT      C>A     ACT
#> C>A_CCA C>A_CCA      C>A     CCA
#> C>A_CCC C>A_CCC      C>A     CCC
#> 7           ...      ...     ... 
#> 
#> **Features: 
#>    mutation
#> 1  C>A_TCA
#> 2  C>A_ACT
#> 3  T>C_ATT
#> 4  C>G_TCT
#> 5  T>G_CTC
#> 6  C>G_CCC
#> 7      ... 
#> 
#> **Types: 
#>  SBS
#>  
#> **Color Variable: 
#>  mutation
#>  
#> **Color Mapping: 
#>  #5ABCEBFF
#>  #050708FF
#>  #D33C32FF
#>  #CBCACBFF
#>  #ABCD72FF
#>  #E7C9C6FF
#>  
#> **Descriptions: 
#>  Single Base Substitution table with one base upstream and downstream
#> 
#> 
#> Slot "sample_annotations":
#>                         Samples Tumor_Subtypes
#> 1: TCGA-94-7557-01A-11D-2122-08           Lung
#> 2: TCGA-56-7582-01A-11D-2042-08           Lung
#> 3: TCGA-77-7335-01A-11D-2042-08           Lung
#> 4: TCGA-97-7938-01A-11D-2167-08           Lung
#> 5: TCGA-EE-A3J5-06A-11D-A20D-08         Breast
#> 6: TCGA-ER-A197-06A-32D-A197-08         Breast
#> 7: TCGA-ER-A19O-06A-11D-A197-08         Breast
#> 
#> 
#> Slot "umap":
#> <0 x 0 matrix>
#> 

auto_predict_grid(musica_annot, "SBS96", cosmic_v2_sigs, "lda")

#> An object of class "musica_result"
#> Slot "signatures":
#>           Signature1   Signature3 Signature4 Signature6 Signature7  Signature13
#> C>A_ACA 1.109833e-02 0.0221723068     0.0365     0.0017     0.0004 3.347572e-04
#> C>A_ACC 9.149341e-03 0.0178716754     0.0309     0.0028     0.0005 6.487361e-04
#> C>A_ACG 1.490070e-03 0.0021383396     0.0183     0.0005     0.0000 3.814459e-05
#> C>A_ACT 6.233885e-03 0.0162651456     0.0243     0.0019     0.0004 8.466585e-04
#> C>A_CCA 6.595870e-03 0.0187817256     0.0461     0.0101     0.0012 1.710090e-03
#> C>A_CCC 7.342368e-03 0.0157604578     0.0614     0.0241     0.0006 1.159257e-03
#> C>A_CCG 8.928404e-04 0.0019633898     0.0088     0.0091     0.0000 2.441665e-04
#> C>A_CCT 7.186582e-03 0.0147228611     0.0432     0.0571     0.0013 1.256768e-03
#> C>A_GCA 8.232604e-03 0.0096965397     0.0376     0.0024     0.0003 1.321096e-04
#> C>A_GCC 5.758021e-03 0.0108433411     0.0399     0.0058     0.0001 7.542440e-04
#> C>A_GCG 6.163352e-04 0.0009291405     0.0227     0.0021     0.0000 9.682078e-05
#> C>A_GCT 4.459080e-03 0.0122153826     0.0258     0.0087     0.0001 5.511037e-04
#> C>A_TCA 1.225006e-02 0.0116532249     0.0330     0.0017     0.0010 4.816605e-02
#> C>A_TCC 1.116223e-02 0.0166067750     0.0538     0.0029     0.0020 1.732956e-02
#> C>A_TCG 2.275496e-03 0.0013572394     0.0104     0.0011     0.0002 2.293157e-03
#> C>A_TCT 1.525910e-02 0.0163280760     0.0370     0.0058     0.0013 1.863458e-02
#> C>G_ACA 1.801068e-03 0.0240026152     0.0097     0.0013     0.0000 3.775165e-03
#> C>G_ACC 2.580909e-03 0.0121603037     0.0054     0.0012     0.0000 9.208248e-04
#> C>G_ACG 5.925480e-04 0.0052754195     0.0031     0.0000     0.0000 1.989049e-05
#> C>G_ACT 2.963986e-03 0.0232776563     0.0054     0.0018     0.0001 3.860632e-03
#> C>G_CCA 1.284983e-03 0.0168325720     0.0105     0.0000     0.0001 5.333651e-03
#> C>G_CCC 7.021348e-04 0.0135314415     0.0097     0.0000     0.0004 9.123035e-04
#> C>G_CCG 5.062896e-04 0.0041764575     0.0063     0.0000     0.0006 2.854673e-04
#> C>G_CCT 1.381543e-03 0.0240463904     0.0094     0.0002     0.0003 6.383098e-03
#> C>G_GCA 6.021227e-04 0.0119168446     0.0070     0.0000     0.0000 1.349420e-03
#> C>G_GCC 2.393352e-03 0.0098236534     0.0091     0.0030     0.0004 7.888077e-04
#> C>G_GCG 2.485340e-07 0.0016710543     0.0062     0.0000     0.0000 8.179121e-05
#> C>G_GCT 8.900807e-04 0.0179143343     0.0060     0.0017     0.0000 2.680123e-03
#> C>G_TCA 1.874853e-03 0.0160414054     0.0032     0.0000     0.0002 2.802349e-01
#> C>G_TCC 2.067419e-03 0.0201499204     0.0105     0.0002     0.0010 6.388529e-02
#> C>G_TCG 3.048970e-04 0.0025279109     0.0031     0.0000     0.0003 9.528486e-03
#> C>G_TCT 3.151574e-03 0.0326736408     0.0050     0.0001     0.0007 3.302255e-01
#> C>T_ACA 2.951453e-02 0.0178721707     0.0120     0.0312     0.0000 1.480181e-03
#> C>T_ACC 1.432275e-02 0.0088960343     0.0075     0.0163     0.0197 4.050871e-05
#> C>T_ACG 1.716469e-01 0.0035727083     0.0028     0.0908     0.0001 3.165947e-04
#> C>T_ACT 1.262376e-02 0.0147976122     0.0059     0.0149     0.0043 0.000000e+00
#> C>T_CCA 2.089645e-02 0.0143950595     0.0210     0.0085     0.0754 3.647278e-03
#> C>T_CCC 1.850170e-02 0.0085447812     0.0144     0.0099     0.1007 1.004229e-03
#> C>T_CCG 9.557722e-02 0.0035184658     0.0076     0.0901     0.0208 1.292021e-04
#> C>T_CCT 1.711331e-02 0.0160755457     0.0201     0.0087     0.0788 6.117700e-04
#> C>T_GCA 2.494381e-02 0.0161272733     0.0087     0.0653     0.0000 2.934138e-03
#> C>T_GCC 2.716149e-02 0.0082086324     0.0080     0.0773     0.0210 2.515171e-03
#> C>T_GCG 1.035708e-01 0.0012129229     0.0023     0.1339     0.0002 3.205265e-03
#> C>T_GCT 1.768985e-02 0.0106116492     0.0082     0.0524     0.0161 1.687731e-03
#> C>T_TCA 1.449210e-02 0.0088802091     0.0035     0.0074     0.1202 1.138421e-01
#> C>T_TCC 1.768078e-02 0.0135295728     0.0070     0.0067     0.2887 1.502484e-02
#> C>T_TCG 7.600222e-02 0.0017054089     0.0011     0.0391     0.0992 6.102098e-03
#> C>T_TCT 1.376170e-02 0.0103044169     0.0077     0.0047     0.0844 2.814066e-02
#> T>A_ATA 4.021520e-03 0.0084285636     0.0048     0.0006     0.0010 1.243618e-03
#> T>A_ATC 2.371144e-03 0.0073730967     0.0039     0.0033     0.0008 1.060002e-03
#> T>A_ATG 2.810910e-03 0.0073573000     0.0100     0.0000     0.0009 1.025544e-04
#> T>A_ATT 8.360909e-03 0.0087539390     0.0030     0.0053     0.0035 9.718121e-04
#> T>A_CTA 1.182587e-03 0.0075713203     0.0075     0.0001     0.0000 2.851590e-04
#> T>A_CTC 1.903167e-03 0.0127254635     0.0111     0.0026     0.0014 2.222873e-04
#> T>A_CTG 1.487961e-03 0.0115085619     0.0342     0.0008     0.0007 2.869518e-04
#> T>A_CTT 2.179344e-03 0.0164561777     0.0115     0.0011     0.0018 8.004460e-07
#> T>A_GTA 6.892894e-04 0.0044347826     0.0069     0.0000     0.0000 1.379604e-04
#> T>A_GTC 5.524095e-04 0.0056154324     0.0052     0.0028     0.0000 2.528674e-04
#> T>A_GTG 1.200229e-03 0.0080701365     0.0133     0.0006     0.0006 1.535175e-04
#> T>A_GTT 2.107137e-03 0.0086791217     0.0045     0.0021     0.0010 1.227456e-04
#> T>A_TTA 5.600155e-03 0.0071326814     0.0045     0.0002     0.0010 4.032613e-04
#> T>A_TTC 1.999079e-03 0.0091027978     0.0046     0.0008     0.0015 7.718133e-04
#> T>A_TTG 1.090066e-03 0.0065658630     0.0082     0.0000     0.0010 2.451942e-04
#> T>A_TTT 3.981023e-03 0.0147121354     0.0045     0.0007     0.0050 1.603460e-04
#> T>C_ATA 1.391577e-02 0.0130357094     0.0084     0.0075     0.0005 5.494542e-04
#> T>C_ATC 6.274961e-03 0.0091862897     0.0020     0.0056     0.0001 1.455067e-04
#> T>C_ATG 1.013764e-02 0.0117169201     0.0081     0.0217     0.0007 2.791780e-04
#> T>C_ATT 9.256316e-03 0.0169788708     0.0036     0.0023     0.0011 2.674614e-04
#> T>C_CTA 4.176675e-03 0.0078950126     0.0052     0.0062     0.0010 4.275147e-05
#> T>C_CTC 5.252593e-03 0.0144311704     0.0026     0.0040     0.0008 4.330754e-05
#> T>C_CTG 7.013225e-03 0.0084229721     0.0100     0.0270     0.0010 5.427253e-05
#> T>C_CTT 6.713813e-03 0.0119324305     0.0054     0.0033     0.0045 7.639440e-07
#> T>C_GTA 1.124784e-02 0.0068500556     0.0061     0.0122     0.0002 1.054081e-03
#> T>C_GTC 6.999724e-03 0.0062607490     0.0016     0.0059     0.0001 5.298997e-04
#> T>C_GTG 4.977593e-03 0.0060987630     0.0042     0.0115     0.0003 6.300957e-04
#> T>C_GTT 1.066741e-02 0.0075093439     0.0024     0.0042     0.0081 6.761879e-04
#> T>C_TTA 8.073616e-03 0.0091154951     0.0028     0.0059     0.0023 7.840561e-04
#> T>C_TTC 4.857381e-03 0.0109537909     0.0016     0.0035     0.0018 1.329749e-04
#> T>C_TTG 8.325454e-03 0.0061130332     0.0036     0.0106     0.0019 3.740240e-05
#> T>C_TTT 6.257106e-03 0.0107743363     0.0022     0.0029     0.0024 1.797028e-04
#> T>G_ATA 1.587636e-03 0.0023513681     0.0000     0.0000     0.0000 3.861998e-04
#> T>G_ATC 1.784091e-03 0.0014642311     0.0002     0.0017     0.0000 1.123832e-04
#> T>G_ATG 1.385831e-03 0.0090537776     0.0015     0.0007     0.0009 3.600070e-05
#> T>G_ATT 3.158539e-03 0.0070309764     0.0002     0.0029     0.0000 1.924938e-04
#> T>G_CTA 3.026912e-04 0.0019741811     0.0000     0.0001     0.0000 1.509421e-05
#> T>G_CTC 2.098502e-03 0.0058242955     0.0013     0.0040     0.0008 1.476248e-04
#> T>G_CTG 1.599549e-03 0.0104646545     0.0046     0.0050     0.0009 4.598730e-05
#> T>G_CTT 2.758538e-03 0.0087243873     0.0012     0.0086     0.0013 4.637147e-04
#> T>G_GTA 9.904500e-05 0.0041444880     0.0000     0.0000     0.0000 1.848986e-05
#> T>G_GTC 2.023656e-04 0.0045019853     0.0000     0.0016     0.0000 9.337351e-05
#> T>G_GTG 1.188353e-03 0.0163914526     0.0018     0.0010     0.0017 1.057919e-05
#> T>G_GTT 8.007233e-04 0.0070672366     0.0002     0.0035     0.0009 5.374121e-05
#> T>G_TTA 1.397554e-03 0.0054271842     0.0000     0.0009     0.0000 5.465818e-04
#> T>G_TTC 1.291737e-03 0.0061602504     0.0003     0.0019     0.0010 2.353471e-04
#> T>G_TTG 2.031077e-03 0.0110765263     0.0030     0.0011     0.0010 4.790220e-07
#> T>G_TTT 4.030128e-03 0.0130009842     0.0011     0.0072     0.0014 6.705883e-04
#>          Signature18 Signature21  Signature24 Signature30
#> C>A_ACA 5.053642e-02      0.0001 2.864599e-02   0.0000000
#> C>A_ACC 1.093982e-02      0.0007 2.021464e-02   0.0000000
#> C>A_ACG 2.288073e-03      0.0000 2.047900e-02   0.0019673
#> C>A_ACT 1.942409e-02      0.0006 2.460015e-02   0.0000000
#> C>A_CCA 8.876811e-02      0.0020 6.355928e-02   0.0000000
#> C>A_CCC 2.064139e-02      0.0014 3.375700e-02   0.0000000
#> C>A_CCG 1.717840e-02      0.0027 2.242899e-02   0.0022624
#> C>A_CCT 3.767696e-02      0.0056 2.008652e-02   0.0000000
#> C>A_GCA 1.287242e-01      0.0001 5.467645e-02   0.0088530
#> C>A_GCC 1.609288e-02      0.0038 1.099460e-01   0.0093449
#> C>A_GCG 9.271417e-03      0.0003 4.652487e-02   0.0008853
#> C>A_GCT 7.201707e-02      0.0023 5.042871e-02   0.0081645
#> C>A_TCA 6.708064e-02      0.0001 3.145126e-02   0.0000000
#> C>A_TCC 4.186901e-02      0.0005 6.253829e-02   0.0000000
#> C>A_TCG 1.516238e-02      0.0000 1.280259e-02   0.0016722
#> C>A_TCT 1.178778e-01      0.0004 3.029768e-02   0.0000000
#> C>G_ACA 1.516888e-03      0.0005 1.199304e-02   0.0000000
#> C>G_ACC 2.498785e-03      0.0008 8.425120e-03   0.0000000
#> C>G_ACG 2.614509e-03      0.0000 0.000000e+00   0.0048200
#> C>G_ACT 3.983011e-03      0.0018 3.881383e-03   0.0000000
#> C>G_CCA 4.584638e-04      0.0001 1.347682e-02   0.0000000
#> C>G_CCC 3.654228e-03      0.0000 1.610569e-02   0.0000000
#> C>G_CCG 5.405774e-03      0.0000 7.974720e-04   0.0032461
#> C>G_CCT 5.519605e-03      0.0003 1.038867e-02   0.0000000
#> C>G_GCA 9.238797e-04      0.0004 4.945061e-03   0.0073775
#> C>G_GCC 6.486479e-04      0.0042 9.369876e-03   0.0065906
#> C>G_GCG 5.428400e-04      0.0000 5.625727e-03   0.0015739
#> C>G_GCT 1.376725e-03      0.0004 3.055770e-03   0.0000000
#> C>G_TCA 1.802969e-03      0.0002 1.022086e-02   0.0000000
#> C>G_TCC 7.863007e-03      0.0003 7.578365e-03   0.0000000
#> C>G_TCG 1.513223e-03      0.0000 1.592714e-03   0.0019673
#> C>G_TCT 3.567102e-03      0.0003 6.100727e-03   0.0000000
#> C>T_ACA 3.825624e-03      0.0051 6.315518e-03   0.0651190
#> C>T_ACC 2.465214e-03      0.0062 3.844327e-03   0.0543970
#> C>T_ACG 8.463957e-03      0.0256 3.110664e-03   0.0204604
#> C>T_ACT 5.605740e-03      0.0030 3.420419e-03   0.0219359
#> C>T_CCA 9.767511e-03      0.0034 6.035447e-03   0.0694472
#> C>T_CCC 5.654828e-03      0.0091 0.000000e+00   0.0638403
#> C>T_CCG 1.618322e-02      0.0318 2.924419e-03   0.0173126
#> C>T_CCT 5.153319e-03      0.0040 3.971202e-03   0.0342318
#> C>T_GCA 2.703069e-03      0.0163 2.367905e-02   0.0485934
#> C>T_GCC 4.199308e-03      0.0392 1.591920e-02   0.0494787
#> C>T_GCG 1.189736e-02      0.0019 1.264808e-02   0.0157387
#> C>T_GCT 1.913381e-03      0.0166 1.812978e-02   0.0188865
#> C>T_TCA 1.492339e-02      0.0064 2.657189e-02   0.0849892
#> C>T_TCC 9.962744e-03      0.0086 1.205569e-02   0.0903010
#> C>T_TCG 1.154972e-02      0.0126 5.363360e-03   0.0151485
#> C>T_TCT 5.398014e-03      0.0064 5.532499e-03   0.0455440
#> T>A_ATA 3.037211e-03      0.0083 1.166747e-03   0.0075743
#> T>A_ATC 8.111672e-04      0.0292 1.517306e-05   0.0037380
#> T>A_ATG 4.503207e-03      0.0027 6.332456e-03   0.0065906
#> T>A_ATT 2.736523e-03      0.0043 4.967216e-06   0.0093447
#> T>A_CTA 4.123464e-04      0.0000 2.766635e-03   0.0053118
#> T>A_CTC 2.491411e-03      0.0037 2.694462e-03   0.0067873
#> T>A_CTG 5.523046e-03      0.0025 6.911165e-03   0.0082628
#> T>A_CTT 3.313713e-03      0.0010 2.915200e-03   0.0000000
#> T>A_GTA 1.875434e-03      0.0000 2.225127e-03   0.0037380
#> T>A_GTC 1.510310e-03      0.0142 4.853149e-03   0.0033445
#> T>A_GTG 1.505369e-03      0.0031 1.957586e-04   0.0050167
#> T>A_GTT 3.611025e-03      0.0041 0.000000e+00   0.0063939
#> T>A_TTA 4.663278e-03      0.0027 0.000000e+00   0.0076726
#> T>A_TTC 2.578156e-03      0.0045 7.994129e-03   0.0064922
#> T>A_TTG 2.016738e-03      0.0000 1.120214e-02   0.0036396
#> T>A_TTT 2.818679e-03      0.0008 2.705788e-04   0.0000000
#> T>C_ATA 2.933054e-03      0.0265 4.726009e-03   0.0000000
#> T>C_ATC 1.513285e-03      0.0165 6.279493e-04   0.0087547
#> T>C_ATG 3.329669e-03      0.0224 2.723759e-03   0.0090498
#> T>C_ATT 5.491318e-03      0.0189 7.122368e-03   0.0000000
#> T>C_CTA 1.921693e-04      0.0219 1.848877e-03   0.0000000
#> T>C_CTC 3.050193e-03      0.0100 4.453690e-03   0.0000000
#> T>C_CTG 2.596464e-03      0.0211 0.000000e+00   0.0000000
#> T>C_CTT 1.310029e-03      0.0204 1.597389e-03   0.0000000
#> T>C_GTA 5.291265e-04      0.1426 1.269750e-03   0.0000000
#> T>C_GTC 2.279131e-03      0.0665 6.710340e-03   0.0065906
#> T>C_GTG 5.871057e-03      0.0884 2.776850e-04   0.0078694
#> T>C_GTT 4.771566e-03      0.1003 4.238478e-03   0.0084596
#> T>C_TTA 3.023036e-04      0.0389 1.682291e-03   0.0000000
#> T>C_TTC 6.260791e-03      0.0313 2.344528e-03   0.0091481
#> T>C_TTG 1.541265e-03      0.0249 4.812742e-04   0.0060004
#> T>C_TTT 1.357797e-03      0.0274 3.581480e-04   0.0000000
#> T>G_ATA 3.861382e-04      0.0000 0.000000e+00   0.0053118
#> T>G_ATC 1.254982e-03      0.0031 1.463951e-03   0.0024592
#> T>G_ATG 2.369216e-03      0.0016 8.476125e-04   0.0062955
#> T>G_ATT 3.386029e-04      0.0018 1.075455e-03   0.0095416
#> T>G_CTA 6.100255e-05      0.0000 0.000000e+00   0.0046233
#> T>G_CTC 2.015697e-03      0.0005 2.185437e-03   0.0060004
#> T>G_CTG 5.311967e-03      0.0004 1.139097e-04   0.0073775
#> T>G_CTT 2.341629e-03      0.0032 2.533718e-03   0.0091481
#> T>G_GTA 1.341532e-03      0.0000 0.000000e+00   0.0032461
#> T>G_GTC 9.204310e-07      0.0006 1.073632e-03   0.0018690
#> T>G_GTG 9.695000e-03      0.0004 2.924697e-03   0.0033445
#> T>G_GTT 2.890554e-03      0.0008 8.250983e-04   0.0056069
#> T>G_TTA 3.631840e-05      0.0000 0.000000e+00   0.0086563
#> T>G_TTC 3.233838e-03      0.0018 0.000000e+00   0.0043282
#> T>G_TTG 7.546018e-04      0.0003 1.360505e-03   0.0082628
#> T>G_TTT 2.126442e-03      0.0003 6.951578e-05   0.0000000
#> 
#> Slot "exposures":
#>             TCGA-56-7582-01A-11D-2042-08 TCGA-77-7335-01A-11D-2042-08
#> Signature1                    38.0547117                 2.951515e+01
#> Signature3                    31.1381600                 4.247401e+01
#> Signature4                    42.4966853                 7.435720e+01
#> Signature6                     9.9390431                 2.892621e+01
#> Signature7                     9.4448542                 1.904947e+01
#> Signature13                    4.4926786                 5.091633e+01
#> Signature18                   11.8085959                 1.304692e-02
#> Signature21                    0.0135092                 1.664935e-02
#> Signature24                   16.4973150                 2.333150e-08
#> Signature30                   28.1144470                 2.873193e+01
#>             TCGA-94-7557-01A-11D-2122-08 TCGA-97-7938-01A-11D-2167-08
#> Signature1                  1.670147e+01                 6.070861e+00
#> Signature3                  1.272421e+01                 1.701443e+00
#> Signature4                  6.985214e+00                 6.258025e+01
#> Signature6                  1.716325e+01                 1.285449e+00
#> Signature7                  2.003852e+01                 7.257755e+00
#> Signature13                 3.200873e+01                 1.282879e+01
#> Signature18                 1.789642e+00                 1.696400e+01
#> Signature21                 4.768777e+00                 1.744839e-02
#> Signature24                 2.208911e-08                 2.619871e-08
#> Signature30                 4.820187e+00                 8.294011e+00
#>             TCGA-EE-A3J5-06A-11D-A20D-08 TCGA-ER-A197-06A-32D-A197-08
#> Signature1                  2.237774e+00                 1.265206e+00
#> Signature3                  6.113964e-01                 3.158933e-01
#> Signature4                  2.518377e+00                 1.909507e+00
#> Signature6                  1.196684e-01                 1.763549e-01
#> Signature7                  1.002094e+02                 1.567182e+00
#> Signature13                 6.100717e-03                 1.776477e-03
#> Signature18                 1.634477e-02                 1.124801e-02
#> Signature21                 6.838518e+00                 3.538011e+00
#> Signature24                 1.552313e-08                 1.810983e-08
#> Signature30                 9.442458e+00                 2.214822e+00
#>             TCGA-ER-A19O-06A-11D-A197-08
#> Signature1                  3.123445e+00
#> Signature3                  1.389646e+00
#> Signature4                  6.165408e-01
#> Signature6                  3.090099e-01
#> Signature7                  4.006470e+01
#> Signature13                 6.701291e-03
#> Signature18                 4.137506e-03
#> Signature21                 1.346391e-02
#> Signature24                 1.239909e-08
#> Signature30                 5.472353e+00
#> 
#> Slot "table_name":
#> [1] "SBS96"
#> 
#> Slot "algorithm":
#> [1] "posterior_LDA"
#> 
#> Slot "musica":
#> An object of class "musica"
#> Slot "variants":
#>        chr     start       end ref alt                       sample
#>   1:  chr1  11020563  11020563   C   A TCGA-94-7557-01A-11D-2122-08
#>   2:  chr1  43430030  43430030   G   T TCGA-94-7557-01A-11D-2122-08
#>   3:  chr1  58682403  58682403   A   G TCGA-94-7557-01A-11D-2122-08
#>   4:  chr1 109508295 109508295   C   G TCGA-94-7557-01A-11D-2122-08
#>   5:  chr1 156384826 156384826   A   C TCGA-94-7557-01A-11D-2122-08
#>  ---                                                               
#> 907: chr19  54885292  54885292   G   A TCGA-ER-A19O-06A-11D-A197-08
#> 908: chr20  49374328  49374328   A   T TCGA-ER-A19O-06A-11D-A197-08
#> 909:  chrX  73213768  73213768   T   G TCGA-ER-A19O-06A-11D-A197-08
#> 910:  chrX 101292834 101292834   G   A TCGA-ER-A19O-06A-11D-A197-08
#> 911:  chrX 107526690 107526690   C   T TCGA-ER-A19O-06A-11D-A197-08
#>      Variant_Type
#>   1:          SBS
#>   2:          SBS
#>   3:          SBS
#>   4:          SBS
#>   5:          SBS
#>  ---             
#> 907:          SBS
#> 908:          SBS
#> 909:          SBS
#> 910:          SBS
#> 911:          SBS
#> 
#> Slot "count_tables":
#> $SBS96
#> Count_Table:  SBS96 
#> Motifs: 96 
#> Samples: 7 
#>  
#> **Annotations: 
#>            motif mutation context
#> C>A_ACA C>A_ACA      C>A     ACA
#> C>A_ACC C>A_ACC      C>A     ACC
#> C>A_ACG C>A_ACG      C>A     ACG
#> C>A_ACT C>A_ACT      C>A     ACT
#> C>A_CCA C>A_CCA      C>A     CCA
#> C>A_CCC C>A_CCC      C>A     CCC
#> 7           ...      ...     ... 
#> 
#> **Features: 
#>    mutation
#> 1  C>A_TCA
#> 2  C>A_ACT
#> 3  T>C_ATT
#> 4  C>G_TCT
#> 5  T>G_CTC
#> 6  C>G_CCC
#> 7      ... 
#> 
#> **Types: 
#>  SBS
#>  
#> **Color Variable: 
#>  mutation
#>  
#> **Color Mapping: 
#>  #5ABCEBFF
#>  #050708FF
#>  #D33C32FF
#>  #CBCACBFF
#>  #ABCD72FF
#>  #E7C9C6FF
#>  
#> **Descriptions: 
#>  Single Base Substitution table with one base upstream and downstream
#> 
#> 
#> Slot "sample_annotations":
#>                         Samples Tumor_Subtypes
#> 1: TCGA-94-7557-01A-11D-2122-08           Lung
#> 2: TCGA-56-7582-01A-11D-2042-08           Lung
#> 3: TCGA-77-7335-01A-11D-2042-08           Lung
#> 4: TCGA-97-7938-01A-11D-2167-08           Lung
#> 5: TCGA-EE-A3J5-06A-11D-A20D-08         Breast
#> 6: TCGA-ER-A197-06A-32D-A197-08         Breast
#> 7: TCGA-ER-A19O-06A-11D-A197-08         Breast
#> 
#> 
#> Slot "umap":
#> <0 x 0 matrix>
#>