Create a UMAP from a musica result

Proportional sample exposures will be used as input into the umap function to generate a two dimensional UMAP.

create_umap(result, n_neighbors = 30, min_dist = 0.75, spread = 1)

Arguments

result
A musica_result object generated by a mutational discovery or prediction tool.

n_neighbors
The size of local neighborhood used for views of manifold approximation. Larger values result in more global the manifold, while smaller values result in more local data being preserved. If n_neighbors is larger than the number of samples, then n_neighbors will automatically be set to the number of samples in the musica_result. Default 30.

min_dist
The effective minimum distance between embedded points. Smaller values will result in a more clustered/clumped embedding where nearby points on the manifold are drawn closer together, while larger values will result on a more even dispersal of points. Default 0.2.

spread
The effective scale of embedded points. In combination with ‘min_dist’, this determines how clustered/clumped the embedded points are. Default 1.

Value

A musica_result object with a new UMAP stored in the UMAP slot.

See also

Examples

data(res_annot)
create_umap(result = res_annot)
#> The parameter 'n_neighbors' cannot be bigger than the total number of samples. Setting 'n_neighbors' to 7.