Discover mutational signatures

Mutational signatures and exposures will be discovered using methods such as Latent Dirichlet Allocation (lda) or Non-Negative Matrix Factorization (nmf). These algorithms will deconvolute a matrix of counts for mutation types in each sample to two matrices: 1) a "signature" matrix containing the probability of each mutation type in each sample and 2) an "exposure" matrix containing the estimated counts for each signature in each sample. Before mutational discovery can be performed, variants from samples first need to be stored in a musica object using the create_musica function and mutation count tables need to be created using functions such as build_standard_table.

discover_signatures(
  musica,
  table_name,
  num_signatures,
  algorithm = "lda",
  seed = 1,
  nstart = 10,
  par_cores = 1
)

Arguments

musica
A musica object.

table_name
Name of the table to use for signature discovery. Needs to be the same name supplied to the table building functions such as build_standard_table.

num_signatures
Number of signatures to discover.

algorithm
Method to use for mutational signature discovery. One of "lda" or "nmf". Default "lda".

seed
Seed to be used for the random number generators in the signature discovery algorithms. Default 1.

nstart
Number of independent random starts used in the mutational signature algorithms. Default 10.

par_cores
Number of parallel cores to use. Only used if method = "nmf". Default 1.

Value

Returns a A musica_result object containing signatures and exposures.

Examples

data(musica)
g <- select_genome("19")
build_standard_table(musica, g, "SBS96", overwrite = TRUE)

#> Building count table from SBS with SBS96 schema

#> Warning: Overwriting counts table: SBS96

discover_signatures(musica = musica, table_name = "SBS96",
num_signatures = 3, algorithm = "lda", seed = 12345, nstart = 1)

#> An object of class "musica_result"
#> Slot "signatures":
#>           Signature1   Signature2   Signature3
#> C>A_ACA 8.565633e-07 1.111993e-02 1.970443e-02
#> C>A_ACC 1.170396e-09 1.334444e-02 1.477833e-02
#> C>A_ACG 1.734163e-11 6.672222e-03 6.834717e-59
#> C>A_ACT 5.559457e-09 6.672219e-03 1.477833e-02
#> C>A_CCA 5.101718e-03 2.183950e-02 3.448276e-02
#> C>A_CCC 2.045925e-11 3.113703e-02 1.477833e-02
#> C>A_CCG 6.578572e-13 2.446481e-02 6.947472e-55
#> C>A_CCT 6.715738e-14 2.001666e-02 9.852217e-03
#> C>A_GCA 5.636593e-03 1.934018e-02 3.940887e-02
#> C>A_GCC 3.557662e-13 1.334444e-02 2.463054e-02
#> C>A_GCG 3.717609e-65 7.257405e-60 1.970443e-02
#> C>A_GCT 1.429697e-09 8.896295e-03 1.477833e-02
#> C>A_TCA 3.969575e-07 2.446461e-02 9.852217e-03
#> C>A_TCC 3.464136e-13 2.224074e-02 1.970443e-02
#> C>A_TCG 2.495100e-13 1.112037e-02 9.852217e-03
#> C>A_TCT 9.713682e-10 2.224074e-02 1.970443e-02
#> C>G_ACA 3.720076e-44 3.720076e-44 3.720076e-44
#> C>G_ACC 5.328026e-03 3.930449e-03 1.477833e-02
#> C>G_ACG 2.974271e-12 4.448148e-03 4.926108e-03
#> C>G_ACT 8.724185e-12 2.224074e-03 9.083644e-58
#> C>G_CCA 1.772080e-11 1.112037e-02 9.852217e-03
#> C>G_CCC 6.420143e-09 1.112037e-02 2.005127e-53
#> C>G_CCG 3.132700e-66 5.403700e-60 9.852217e-03
#> C>G_CCT 1.854540e-13 4.448148e-03 9.852217e-03
#> C>G_GCA 7.805780e-12 6.672222e-03 9.852217e-03
#> C>G_GCC 1.028020e-11 4.448148e-03 9.852217e-03
#> C>G_GCG 9.901391e-08 6.672171e-03 4.612018e-51
#> C>G_GCT 9.384787e-09 8.896291e-03 3.956190e-52
#> C>G_TCA 1.417124e-06 6.894556e-02 1.477833e-02
#> C>G_TCC 2.727072e-08 2.668887e-02 4.926108e-03
#> C>G_TCG 4.562089e-12 8.896295e-03 2.974574e-57
#> C>G_TCT 6.788051e-07 6.672187e-02 1.477833e-02
#> C>T_ACA 6.143451e-13 6.672222e-03 3.448276e-02
#> C>T_ACC 1.375258e-02 6.267443e-03 1.477833e-02
#> C>T_ACG 5.789105e-03 1.926170e-02 3.940887e-02
#> C>T_ACT 2.477903e-08 8.896283e-03 4.926108e-03
#> C>T_CCA 6.252214e-02 2.342831e-02 3.448276e-02
#> C>T_CCC 1.271463e-01 1.018974e-02 1.477833e-02
#> C>T_CCG 2.951043e-02 3.374372e-02 3.940887e-02
#> C>T_CCT 8.264461e-02 1.307340e-02 2.463054e-02
#> C>T_GCA 5.277229e-08 1.556849e-02 2.463054e-02
#> C>T_GCC 2.916937e-02 1.835072e-02 2.955665e-02
#> C>T_GCG 6.122445e-03 3.021053e-02 1.477833e-02
#> C>T_GCT 1.088506e-02 1.219120e-02 2.463054e-02
#> C>T_TCA 9.568696e-02 3.082670e-02 2.463054e-02
#> C>T_TCC 2.323759e-01 2.744748e-03 2.463054e-02
#> C>T_TCG 1.060298e-01 2.105616e-02 3.448276e-02
#> C>T_TCT 3.876200e-02 3.565514e-02 1.970443e-02
#> T>A_ATA 2.468900e-03 9.535922e-04 3.828059e-49
#> T>A_ATC 5.047330e-03 4.074894e-03 4.926108e-03
#> T>A_ATG 8.967563e-12 4.448148e-03 1.970443e-02
#> T>A_ATT 4.810052e-03 4.196996e-03 4.926108e-03
#> T>A_CTA 2.672073e-12 8.896295e-03 4.926108e-03
#> T>A_CTC 1.695764e-08 6.672213e-03 4.926108e-03
#> T>A_CTG 5.155128e-03 2.181201e-02 9.852217e-03
#> T>A_CTT 1.001879e-11 6.672222e-03 1.477833e-02
#> T>A_GTA 8.643989e-03 2.270236e-67 3.102942e-61
#> T>A_GTC 2.207210e-12 2.224074e-03 5.397067e-58
#> T>A_GTG 5.328529e-03 3.930191e-03 1.710565e-54
#> T>A_GTT 8.459211e-08 6.672178e-03 9.852217e-03
#> T>A_TTA 5.962197e-03 5.828183e-03 2.058781e-50
#> T>A_TTC 1.958204e-65 3.808540e-60 4.926108e-03
#> T>A_TTG 4.104842e-14 6.672222e-03 1.565835e-55
#> T>A_TTT 5.339033e-03 1.700712e-03 4.926108e-03
#> T>C_ATA 2.165275e-05 1.110923e-02 6.626842e-51
#> T>C_ATC 4.810071e-03 4.196987e-03 1.744190e-53
#> T>C_ATG 2.817898e-08 1.556850e-02 2.463054e-02
#> T>C_ATT 1.719898e-02 4.493944e-03 4.926108e-03
#> T>C_CTA 7.395434e-13 4.448148e-03 4.926108e-03
#> T>C_CTC 1.836823e-11 2.224074e-03 9.852217e-03
#> T>C_CTG 3.805789e-08 1.556850e-02 9.852217e-03
#> T>C_CTT 3.227388e-66 2.237098e-59 9.852217e-03
#> T>C_GTA 7.919951e-07 6.671814e-03 4.926108e-03
#> T>C_GTC 2.170894e-07 6.672110e-03 4.926108e-03
#> T>C_GTG 6.917777e-03 3.112373e-03 4.926108e-03
#> T>C_GTT 2.015222e-02 2.974225e-03 4.926108e-03
#> T>C_TTA 6.843888e-03 3.150396e-03 9.852217e-03
#> T>C_TTC 6.260389e-03 5.674735e-03 1.477833e-02
#> T>C_TTG 6.260647e-03 5.674602e-03 4.926108e-03
#> T>C_TTT 1.076425e-02 1.133000e-03 4.926108e-03
#> T>G_ATA 7.069703e-12 4.448148e-03 1.536976e-57
#> T>G_ATC 5.339911e-03 1.700260e-03 2.771505e-53
#> T>G_ATG 4.321995e-03 2.278718e-64 1.968107e-58
#> T>G_ATT 2.143521e-12 6.672222e-03 4.926108e-03
#> T>G_CTA 4.321995e-03 4.833546e-70 6.164008e-65
#> T>G_CTC 4.870697e-08 8.896270e-03 4.926108e-03
#> T>G_CTG 5.339886e-03 1.700273e-03 2.631376e-55
#> T>G_CTT 1.121114e-11 4.448148e-03 2.813161e-58
#> T>G_GTA 2.757935e-12 2.224074e-03 4.926108e-03
#> T>G_GTC 3.720076e-44 3.720076e-44 3.720076e-44
#> T>G_GTG 2.246626e-13 4.448148e-03 1.352918e-55
#> T>G_GTT 2.017090e-65 4.037253e-59 4.926108e-03
#> T>G_TTA 1.056428e-65 1.824919e-60 4.926108e-03
#> T>G_TTC 3.720076e-44 3.720076e-44 3.720076e-44
#> T>G_TTG 1.466616e-06 6.671467e-03 4.373176e-51
#> T>G_TTT 2.221820e-03 1.080738e-03 8.692238e-50
#> 
#> Slot "exposures":
#>            TCGA-56-7582-01A-11D-2042-08 TCGA-77-7335-01A-11D-2042-08
#> Signature1                   0.06265911                  23.71682870
#> Signature2                   0.06265910                 250.22049382
#> Signature3                 191.87468179                   0.06267747
#>            TCGA-94-7557-01A-11D-2122-08 TCGA-97-7938-01A-11D-2167-08
#> Signature1                  25.94050159                   6.13819789
#> Signature2                  90.99687860                 110.79918230
#> Signature3                   0.06261981                   0.06261981
#>            TCGA-EE-A3J5-06A-11D-A20D-08 TCGA-ER-A197-06A-32D-A197-08
#> Signature1                 121.87475221                   0.06166572
#> Signature2                   0.06262389                   0.06166567
#> Signature3                   0.06262390                  10.87666861
#>            TCGA-ER-A19O-06A-11D-A197-08
#> Signature1                  50.87502015
#> Signature2                   0.06248992
#> Signature3                   0.06248993
#> 
#> Slot "table_name":
#> [1] "SBS96"
#> 
#> Slot "algorithm":
#> [1] "LDA"
#> 
#> Slot "musica":
#> An object of class "musica"
#> Slot "variants":
#>        chr     start       end ref alt                       sample
#>   1:  chr1  11020563  11020563   C   A TCGA-94-7557-01A-11D-2122-08
#>   2:  chr1  43430030  43430030   G   T TCGA-94-7557-01A-11D-2122-08
#>   3:  chr1  58682403  58682403   A   G TCGA-94-7557-01A-11D-2122-08
#>   4:  chr1 109508295 109508295   C   G TCGA-94-7557-01A-11D-2122-08
#>   5:  chr1 156384826 156384826   A   C TCGA-94-7557-01A-11D-2122-08
#>  ---                                                               
#> 907: chr19  54885292  54885292   G   A TCGA-ER-A19O-06A-11D-A197-08
#> 908: chr20  49374328  49374328   A   T TCGA-ER-A19O-06A-11D-A197-08
#> 909:  chrX  73213768  73213768   T   G TCGA-ER-A19O-06A-11D-A197-08
#> 910:  chrX 101292834 101292834   G   A TCGA-ER-A19O-06A-11D-A197-08
#> 911:  chrX 107526690 107526690   C   T TCGA-ER-A19O-06A-11D-A197-08
#>      Variant_Type
#>   1:          SBS
#>   2:          SBS
#>   3:          SBS
#>   4:          SBS
#>   5:          SBS
#>  ---             
#> 907:          SBS
#> 908:          SBS
#> 909:          SBS
#> 910:          SBS
#> 911:          SBS
#> 
#> Slot "count_tables":
#> $SBS96
#> Count_Table:  SBS96 
#> Motifs: 96 
#> Samples: 7 
#>  
#> **Annotations: 
#>            motif mutation context
#> C>A_ACA C>A_ACA      C>A     ACA
#> C>A_ACC C>A_ACC      C>A     ACC
#> C>A_ACG C>A_ACG      C>A     ACG
#> C>A_ACT C>A_ACT      C>A     ACT
#> C>A_CCA C>A_CCA      C>A     CCA
#> C>A_CCC C>A_CCC      C>A     CCC
#> 7           ...      ...     ... 
#> 
#> **Features: 
#>    mutation
#> 1  C>A_TCA
#> 2  C>A_ACT
#> 3  T>C_ATT
#> 4  C>G_TCT
#> 5  T>G_CTC
#> 6  C>G_CCC
#> 7      ... 
#> 
#> **Types: 
#>  SBS
#>  
#> **Color Variable: 
#>  mutation
#>  
#> **Color Mapping: 
#>  #5ABCEBFF
#>  #050708FF
#>  #D33C32FF
#>  #CBCACBFF
#>  #ABCD72FF
#>  #E7C9C6FF
#>  
#> **Descriptions: 
#>  Single Base Substitution table with one base upstream and downstream
#> 
#> $SBS96
#> Count_Table:  SBS96 
#> Motifs: 96 
#> Samples: 7 
#>  
#> **Annotations: 
#>            motif mutation context
#> C>A_ACA C>A_ACA      C>A     ACA
#> C>A_ACC C>A_ACC      C>A     ACC
#> C>A_ACG C>A_ACG      C>A     ACG
#> C>A_ACT C>A_ACT      C>A     ACT
#> C>A_CCA C>A_CCA      C>A     CCA
#> C>A_CCC C>A_CCC      C>A     CCC
#> 7           ...      ...     ... 
#> 
#> **Features: 
#>    mutation
#> 1  C>A_CCG
#> 2  C>A_GCA
#> 3  T>C_ATG
#> 4  C>G_TGC
#> 5  T>G_AGC
#> 6  C>G_CCT
#> 7      ... 
#> 
#> **Types: 
#>  SBS
#>  
#> **Color Variable: 
#>  mutation
#>  
#> **Color Mapping: 
#>  #5ABCEBFF
#>  #050708FF
#>  #D33C32FF
#>  #CBCACBFF
#>  #ABCD72FF
#>  #E7C9C6FF
#>  
#> **Descriptions: 
#>  Single Base Substitution table with one base upstream and downstream
#> 
#> 
#> Slot "sample_annotations":
#>                         Samples Tumor_Subtypes
#> 1: TCGA-94-7557-01A-11D-2122-08           Lung
#> 2: TCGA-56-7582-01A-11D-2042-08           Lung
#> 3: TCGA-77-7335-01A-11D-2042-08           Lung
#> 4: TCGA-97-7938-01A-11D-2167-08           Lung
#> 5: TCGA-EE-A3J5-06A-11D-A20D-08         Breast
#> 6: TCGA-ER-A197-06A-32D-A197-08         Breast
#> 7: TCGA-ER-A19O-06A-11D-A197-08         Breast
#> 
#> 
#> Slot "umap":
#> <0 x 0 matrix>
#>