Generate result_grid from musica based on annotation and range of k

Generate result_grid from musica based on annotation and range of k

generate_result_grid(
  musica,
  table_name,
  algorithm = "lda",
  annotation = NA,
  k_start,
  k_end,
  n_start = 1,
  seed = NULL,
  par_cores = FALSE,
  verbose = FALSE
)

Arguments

musica

A musica object.

table_name

Name of table used for signature discovery

algorithm

Algorithm for signature discovery

annotation

Sample annotation to split results into

k_start

Lower range of number of signatures for discovery

k_end

Upper range of number of signatures for discovery

n_start

Number of times to discover signatures and compare based on posterior loglikihood

seed

Seed to use for reproducible results, set to null to disable

par_cores

Number of parallel cores to use (NMF only)

verbose

Whether to output loop iterations

Value

A result object containing signatures and sample weights

Examples

data(musica_sbs96)
grid <- generate_result_grid(musica_sbs96, "SBS96", "lda", k_start = 2, 
k_end = 5)