Visualize clustering results

The clustering results can be visualized on a UMAP panel. Three different types of plots can be generated using this function: cluster-by-signature plot, cluster-by-annotation plot, and a single UMAP plot.

Usage

plot_cluster(
  musica,
  model_name,
  modality = "SBS96",
  result_name = "result",
  clusters,
  group = "signature",
  annotation = NULL,
  plotly = TRUE
)

Arguments

musica

A musica object containing a mutational discovery or prediction. A two-dimensional UMAP has to be stored in this object.

model_name

The name of the desired model.

modality

The modality of the model. Must be "SBS96", "DBS78", or "IND83". Default "SBS96".

result_name

Name of the result list entry containing desired model. Default "result".

clusters

The result generated from cluster_exposure function.

group

A single character string indicating the grouping factor. Possible options are: "signature" (columns are signatures in a grid), "annotation" (columns are sample annotation), and "none" (a single UMAP plot). Default is "signature".

annotation

Column name of annotation.

plotly

If TRUE, the plot will be made interactive using plotly.

Value

Generate a ggplot or plotly object.

See also

Examples

set.seed(123)
data(res_annot)
# Get clustering result
clust_out <- cluster_exposure(
  musica = res_annot, model_name = "res_annot",
  nclust = 2, iter.max = 15
)
#> Metric: 'euclidean'; comparing: 7 vectors.
# UMAP
create_umap(musica = res_annot, model_name = "res_annot")
#> The parameter 'n_neighbors' cannot be bigger than the total number of samples. Setting 'n_neighbors' to 7.
# generate cluster X signature plot
plot_cluster(
  musica = res_annot, model_name = "res_annot",
  clusters = clust_out, group = "signature"
)

# generate cluster X annotation plot
plot_cluster(
  musica = res_annot, model_name = "res_annot",
  clusters = clust_out, group = "annotation",
  annotation = "Tumor_Subtypes"
)

# generate a single UMAP plot
plot_cluster(
  musica = res_annot, model_name = "res_annot",
  clusters = clust_out, group = "none"
)