Prediction of exposures in new samples using pre-existing signatures

Exposures for samples will be predicted using an existing set of signatures stored in a musica_result object. Algorithms available for prediction include a modify version of "lda", "decompTumor2Sig", and "deconstructSigs".

predict_exposure(
  musica,
  g,
  table_name,
  signature_res,
  algorithm = c("lda", "decompTumor2Sig", "deconstructSigs"),
  signatures_to_use = seq_len(ncol(signatures(signature_res))),
  verbose = FALSE
)

Arguments

musica
A musica object.

g
A BSgenome object indicating which genome reference the variants and their coordinates were derived from. Only used if algorithm = "deconstructSigs"

table_name
Name of table used for posterior prediction. Must match the table type used to generate the prediction signatures

signature_res
Signatures used to predict exposures for the samples musica object. Existing signatures need to stored in a musica_result object.

algorithm
Algorithm to use for prediction of exposures. One of "lda", "decompTumor2Sig", or "deconstructSigs".

signatures_to_use
Which signatures in the signature_res result object to use. Default is to use all signatures.

verbose
If TRUE, progress will be printing. Only used if algorithm = "lda". Default FALSE.

Value

Returns a A musica_result object containing signatures given by the signature_res parameter and exposures predicted from these signatures.

Examples

data(musica)
data(cosmic_v2_sigs)
g <- select_genome("19")
build_standard_table(musica, g, "SBS96", overwrite = TRUE)

#> Building count table from SBS with SBS96 schema

#> Warning: Overwriting counts table: SBS96

result <- predict_exposure(musica = musica, table_name = "SBS96",
signature_res = cosmic_v2_sigs, algorithm = "lda")

# Predict using LDA-like algorithm with seed set to 1
set.seed(1)
predict_exposure(musica = musica, table_name = "SBS96",
signature_res = cosmic_v2_sigs, algorithm = "lda")

#> An object of class "musica_result"
#> Slot "signatures":
#>           Signature1   Signature2   Signature3 Signature4   Signature5
#> C>A_ACA 1.109833e-02 6.827082e-04 0.0221723068     0.0365 0.0149415477
#> C>A_ACC 9.149341e-03 6.191072e-04 0.0178716754     0.0309 0.0089609180
#> C>A_ACG 1.490070e-03 9.927896e-05 0.0021383396     0.0183 0.0022078460
#> C>A_ACT 6.233885e-03 3.238914e-04 0.0162651456     0.0243 0.0092069053
#> C>A_CCA 6.595870e-03 6.774450e-04 0.0187817256     0.0461 0.0096749043
#> C>A_CCC 7.342368e-03 2.136810e-04 0.0157604578     0.0614 0.0049523006
#> C>A_CCG 8.928404e-04 6.770460e-06 0.0019633898     0.0088 0.0028006273
#> C>A_CCT 7.186582e-03 4.163329e-04 0.0147228611     0.0432 0.0110134658
#> C>A_GCA 8.232604e-03 3.520134e-04 0.0096965397     0.0376 0.0118921690
#> C>A_GCC 5.758021e-03 1.338169e-04 0.0108433411     0.0399 0.0092478575
#> C>A_GCG 6.163352e-04 1.784417e-04 0.0009291405     0.0227 0.0028091885
#> C>A_GCT 4.459080e-03 1.228320e-04 0.0122153826     0.0258 0.0103012675
#> C>A_TCA 1.225006e-02 1.512743e-02 0.0116532249     0.0330 0.0147740162
#> C>A_TCC 1.116223e-02 6.532492e-03 0.0166067750     0.0538 0.0120434650
#> C>A_TCG 2.275496e-03 1.656455e-03 0.0013572394     0.0104 0.0039023624
#> C>A_TCT 1.525910e-02 1.239461e-02 0.0163280760     0.0370 0.0182433957
#> C>G_ACA 1.801068e-03 2.634810e-04 0.0240026152     0.0097 0.0116710217
#> C>G_ACC 2.580909e-03 2.698660e-04 0.0121603037     0.0054 0.0072920886
#> C>G_ACG 5.925480e-04 2.192339e-04 0.0052754195     0.0031 0.0023038392
#> C>G_ACT 2.963986e-03 6.109735e-04 0.0232776563     0.0054 0.0116962457
#> C>G_CCA 1.284983e-03 2.777208e-05 0.0168325720     0.0105 0.0075375232
#> C>G_CCC 7.021348e-04 2.796439e-04 0.0135314415     0.0097 0.0076332267
#> C>G_CCG 5.062896e-04 1.916158e-05 0.0041764575     0.0063 0.0026137604
#> C>G_CCT 1.381543e-03 3.127816e-04 0.0240463904     0.0094 0.0094170979
#> C>G_GCA 6.021227e-04 4.483851e-05 0.0119168446     0.0070 0.0055594231
#> C>G_GCC 2.393352e-03 1.452010e-05 0.0098236534     0.0091 0.0053888012
#> C>G_GCG 2.485340e-07 4.066589e-05 0.0016710543     0.0062 0.0011005896
#> C>G_GCT 8.900807e-04 2.684295e-04 0.0179143343     0.0060 0.0060412478
#> C>G_TCA 1.874853e-03 3.724185e-02 0.0160414054     0.0032 0.0026810570
#> C>G_TCC 2.067419e-03 1.941341e-05 0.0201499204     0.0105 0.0079240483
#> C>G_TCG 3.048970e-04 1.625465e-03 0.0025279109     0.0031 0.0013190756
#> C>G_TCT 3.151574e-03 6.687990e-02 0.0326736408     0.0050 0.0066445957
#> C>T_ACA 2.951453e-02 7.441557e-03 0.0178721707     0.0120 0.0218391876
#> C>T_ACC 1.432275e-02 2.726312e-03 0.0088960343     0.0075 0.0127561056
#> C>T_ACG 1.716469e-01 3.322083e-03 0.0035727083     0.0028 0.0167601771
#> C>T_ACT 1.262376e-02 3.326528e-03 0.0147976122     0.0059 0.0164779547
#> C>T_CCA 2.089645e-02 1.501951e-02 0.0143950595     0.0210 0.0227692094
#> C>T_CCC 1.850170e-02 3.516918e-03 0.0085447812     0.0144 0.0175094763
#> C>T_CCG 9.557722e-02 4.979276e-03 0.0035184658     0.0076 0.0128622376
#> C>T_CCT 1.711331e-02 8.956553e-03 0.0160755457     0.0201 0.0204295134
#> C>T_GCA 2.494381e-02 6.390808e-03 0.0161272733     0.0087 0.0200382586
#> C>T_GCC 2.716149e-02 1.995817e-03 0.0082086324     0.0080 0.0180223122
#> C>T_GCG 1.035708e-01 3.030210e-04 0.0012129229     0.0023 0.0131938176
#> C>T_GCT 1.768985e-02 3.265818e-03 0.0106116492     0.0082 0.0195029153
#> C>T_TCA 1.449210e-02 4.199414e-01 0.0088802091     0.0035 0.0109975603
#> C>T_TCC 1.768078e-02 8.197250e-02 0.0135295728     0.0070 0.0206449648
#> C>T_TCG 7.600222e-02 4.772019e-02 0.0017054089     0.0011 0.0075344896
#> C>T_TCT 1.376170e-02 2.286749e-01 0.0103044169     0.0077 0.0117874618
#> T>A_ATA 4.021520e-03 1.323770e-07 0.0084285636     0.0048 0.0089029043
#> T>A_ATC 2.371144e-03 1.130696e-04 0.0073730967     0.0039 0.0073992025
#> T>A_ATG 2.810910e-03 5.329294e-04 0.0073573000     0.0100 0.0115080515
#> T>A_ATT 8.360909e-03 1.491110e-04 0.0087539390     0.0030 0.0111936232
#> T>A_CTA 1.182587e-03 1.545990e-04 0.0075713203     0.0075 0.0050431635
#> T>A_CTC 1.903167e-03 4.640190e-04 0.0127254635     0.0111 0.0062088437
#> T>A_CTG 1.487961e-03 2.304098e-04 0.0115085619     0.0342 0.0105075690
#> T>A_CTT 2.179344e-03 5.748856e-04 0.0164561777     0.0115 0.0097846407
#> T>A_GTA 6.892894e-04 1.149942e-04 0.0044347826     0.0069 0.0067403877
#> T>A_GTC 5.524095e-04 2.938621e-04 0.0056154324     0.0052 0.0040215965
#> T>A_GTG 1.200229e-03 8.872134e-05 0.0080701365     0.0133 0.0068143523
#> T>A_GTT 2.107137e-03 2.157555e-04 0.0086791217     0.0045 0.0051010329
#> T>A_TTA 5.600155e-03 8.075249e-05 0.0071326814     0.0045 0.0092056373
#> T>A_TTC 1.999079e-03 4.802898e-06 0.0091027978     0.0046 0.0068353442
#> T>A_TTG 1.090066e-03 6.660504e-05 0.0065658630     0.0082 0.0071442088
#> T>A_TTT 3.981023e-03 2.763159e-04 0.0147121354     0.0045 0.0102408504
#> T>C_ATA 1.391577e-02 1.303510e-03 0.0130357094     0.0084 0.0353667351
#> T>C_ATC 6.274961e-03 4.255767e-04 0.0091862897     0.0020 0.0137711021
#> T>C_ATG 1.013764e-02 5.754730e-04 0.0117169201     0.0081 0.0284490517
#> T>C_ATT 9.256316e-03 1.488164e-03 0.0169788708     0.0036 0.0273029742
#> T>C_CTA 4.176675e-03 5.466192e-04 0.0078950126     0.0052 0.0142386298
#> T>C_CTC 5.252593e-03 3.923326e-04 0.0144311704     0.0026 0.0124895004
#> T>C_CTG 7.013225e-03 3.620245e-04 0.0084229721     0.0100 0.0187416312
#> T>C_CTT 6.713813e-03 5.609001e-04 0.0119324305     0.0054 0.0190540829
#> T>C_GTA 1.124784e-02 4.673852e-06 0.0068500556     0.0061 0.0162949828
#> T>C_GTC 6.999724e-03 1.860659e-04 0.0062607490     0.0016 0.0095745606
#> T>C_GTG 4.977593e-03 4.950440e-07 0.0060987630     0.0042 0.0141039293
#> T>C_GTT 1.066741e-02 5.785981e-04 0.0075093439     0.0024 0.0156668747
#> T>C_TTA 8.073616e-03 1.019324e-04 0.0091154951     0.0028 0.0170596410
#> T>C_TTC 4.857381e-03 4.703437e-04 0.0109537909     0.0016 0.0141957091
#> T>C_TTG 8.325454e-03 1.923542e-04 0.0061130332     0.0036 0.0125965017
#> T>C_TTT 6.257106e-03 5.853303e-04 0.0107743363     0.0022 0.0173751391
#> T>G_ATA 1.587636e-03 3.371841e-05 0.0023513681     0.0000 0.0034621520
#> T>G_ATC 1.784091e-03 2.484483e-05 0.0014642311     0.0002 0.0022465927
#> T>G_ATG 1.385831e-03 2.734541e-04 0.0090537776     0.0015 0.0054902168
#> T>G_ATT 3.158539e-03 2.176949e-04 0.0070309764     0.0002 0.0038212267
#> T>G_CTA 3.026912e-04 1.137846e-04 0.0019741811     0.0000 0.0020416489
#> T>G_CTC 2.098502e-03 2.209509e-05 0.0058242955     0.0013 0.0034785021
#> T>G_CTG 1.599549e-03 2.282459e-04 0.0104646545     0.0046 0.0071474562
#> T>G_CTT 2.758538e-03 6.711134e-05 0.0087243873     0.0012 0.0114868115
#> T>G_GTA 9.904500e-05 9.555239e-05 0.0041444880     0.0000 0.0016276645
#> T>G_GTC 2.023656e-04 4.700238e-05 0.0045019853     0.0000 0.0003277349
#> T>G_GTG 1.188353e-03 1.099257e-04 0.0163914526     0.0018 0.0059488798
#> T>G_GTT 8.007233e-04 8.647718e-05 0.0070672366     0.0002 0.0033074666
#> T>G_TTA 1.397554e-03 7.173695e-05 0.0054271842     0.0000 0.0052028744
#> T>G_TTC 1.291737e-03 1.428146e-05 0.0061602504     0.0003 0.0051316079
#> T>G_TTG 2.031077e-03 2.066152e-04 0.0110765263     0.0030 0.0060552541
#> T>G_TTT 4.030128e-03 2.359820e-05 0.0130009842     0.0011 0.0133699358
#>         Signature6 Signature7   Signature8 Signature9 Signature10 Signature11
#> C>A_ACA     0.0017     0.0004 3.671800e-02     0.0120      0.0007      0.0002
#> C>A_ACC     0.0028     0.0005 3.324572e-02     0.0067      0.0010      0.0010
#> C>A_ACG     0.0005     0.0000 2.525311e-03     0.0005      0.0003      0.0000
#> C>A_ACT     0.0019     0.0004 3.359855e-02     0.0068      0.0092      0.0002
#> C>A_CCA     0.0101     0.0012 3.172376e-02     0.0098      0.0031      0.0007
#> C>A_CCC     0.0241     0.0006 2.550541e-02     0.0057      0.0009      0.0017
#> C>A_CCG     0.0091     0.0000 1.159624e-03     0.0000      0.0007      0.0010
#> C>A_CCT     0.0571     0.0013 2.879117e-02     0.0091      0.0160      0.0014
#> C>A_GCA     0.0024     0.0003 2.368233e-02     0.0118      0.0014      0.0004
#> C>A_GCC     0.0058     0.0001 1.582190e-02     0.0092      0.0022      0.0010
#> C>A_GCG     0.0021     0.0000 8.509138e-04     0.0000      0.0002      0.0000
#> C>A_GCT     0.0087     0.0001 2.106124e-02     0.0085      0.0088      0.0006
#> C>A_TCA     0.0017     0.0010 2.703238e-02     0.0222      0.0374      0.0009
#> C>A_TCC     0.0029     0.0020 1.808977e-02     0.0043      0.0103      0.0025
#> C>A_TCG     0.0011     0.0002 1.694875e-03     0.0000      0.0031      0.0001
#> C>A_TCT     0.0058     0.0013 3.814133e-02     0.0322      0.3083      0.0004
#> C>G_ACA     0.0013     0.0000 8.356845e-03     0.0048      0.0005      0.0007
#> C>G_ACC     0.0012     0.0000 4.306380e-03     0.0023      0.0003      0.0003
#> C>G_ACG     0.0000     0.0000 5.844153e-04     0.0000      0.0000      0.0000
#> C>G_ACT     0.0018     0.0001 8.634907e-03     0.0038      0.0002      0.0009
#> C>G_CCA     0.0000     0.0001 6.619077e-03     0.0018      0.0000      0.0000
#> C>G_CCC     0.0000     0.0004 6.077703e-03     0.0020      0.0000      0.0002
#> C>G_CCG     0.0000     0.0006 6.560334e-04     0.0000      0.0000      0.0008
#> C>G_CCT     0.0002     0.0003 7.805027e-03     0.0039      0.0000      0.0002
#> C>G_GCA     0.0000     0.0000 3.724568e-03     0.0011      0.0000      0.0000
#> C>G_GCC     0.0030     0.0004 3.046597e-03     0.0029      0.0002      0.0002
#> C>G_GCG     0.0000     0.0000 3.212128e-04     0.0000      0.0000      0.0000
#> C>G_GCT     0.0017     0.0000 5.775664e-03     0.0044      0.0004      0.0000
#> C>G_TCA     0.0000     0.0002 4.117879e-03     0.0033      0.0000      0.0001
#> C>G_TCC     0.0002     0.0010 3.799650e-03     0.0025      0.0000      0.0011
#> C>G_TCG     0.0000     0.0003 2.565302e-05     0.0000      0.0000      0.0003
#> C>G_TCT     0.0001     0.0007 6.247554e-03     0.0049      0.0001      0.0008
#> C>T_ACA     0.0312     0.0000 1.806669e-02     0.0093      0.0000      0.0225
#> C>T_ACC     0.0163     0.0197 5.650015e-03     0.0056      0.0032      0.1099
#> C>T_ACG     0.0908     0.0001 1.926511e-02     0.0125      0.0126      0.0072
#> C>T_ACT     0.0149     0.0043 2.080597e-02     0.0076      0.0047      0.0639
#> C>T_CCA     0.0085     0.0754 4.865810e-03     0.0098      0.0012      0.0310
#> C>T_CCC     0.0099     0.1007 3.980150e-03     0.0069      0.0024      0.1518
#> C>T_CCG     0.0901     0.0208 8.338639e-03     0.0076      0.0109      0.0135
#> C>T_CCT     0.0087     0.0788 1.884363e-02     0.0097      0.0043      0.0816
#> C>T_GCA     0.0653     0.0000 2.261263e-03     0.0062      0.0000      0.0261
#> C>T_GCC     0.0773     0.0210 1.616649e-03     0.0069      0.0134      0.0975
#> C>T_GCG     0.1339     0.0002 1.160671e-02     0.0088      0.0270      0.0090
#> C>T_GCT     0.0524     0.0161 6.284929e-03     0.0101      0.0152      0.0522
#> C>T_TCA     0.0074     0.1202 7.180669e-03     0.0050      0.0037      0.0302
#> C>T_TCC     0.0067     0.2887 4.061033e-03     0.0084      0.0211      0.1589
#> C>T_TCG     0.0391     0.0992 5.535122e-03     0.0047      0.2141      0.0080
#> C>T_TCT     0.0047     0.0844 1.220907e-02     0.0096      0.0392      0.0954
#> T>A_ATA     0.0006     0.0010 1.336508e-02     0.0121      0.0000      0.0002
#> T>A_ATC     0.0033     0.0008 1.243120e-02     0.0042      0.0002      0.0001
#> T>A_ATG     0.0000     0.0009 1.403666e-02     0.0068      0.0000      0.0007
#> T>A_ATT     0.0053     0.0035 2.410992e-02     0.0185      0.0012      0.0003
#> T>A_CTA     0.0001     0.0000 1.202422e-02     0.0055      0.0000      0.0004
#> T>A_CTC     0.0026     0.0014 1.788109e-02     0.0049      0.0001      0.0011
#> T>A_CTG     0.0008     0.0007 1.635677e-02     0.0051      0.0003      0.0006
#> T>A_CTT     0.0011     0.0018 2.624780e-02     0.0069      0.0009      0.0004
#> T>A_GTA     0.0000     0.0000 8.139179e-03     0.0030      0.0000      0.0003
#> T>A_GTC     0.0028     0.0000 7.910080e-03     0.0029      0.0001      0.0003
#> T>A_GTG     0.0006     0.0006 9.057263e-03     0.0028      0.0000      0.0004
#> T>A_GTT     0.0021     0.0010 1.781224e-02     0.0036      0.0004      0.0001
#> T>A_TTA     0.0002     0.0010 1.514000e-02     0.0215      0.0000      0.0003
#> T>A_TTC     0.0008     0.0015 1.209355e-02     0.0027      0.0008      0.0000
#> T>A_TTG     0.0000     0.0010 8.362547e-03     0.0140      0.0000      0.0000
#> T>A_TTT     0.0007     0.0050 2.765330e-02     0.0139      0.0028      0.0008
#> T>C_ATA     0.0075     0.0005 1.620651e-02     0.0252      0.0031      0.0006
#> T>C_ATC     0.0056     0.0001 7.787803e-03     0.0105      0.0052      0.0015
#> T>C_ATG     0.0217     0.0007 9.852584e-03     0.0176      0.0026      0.0008
#> T>C_ATT     0.0023     0.0011 2.182946e-02     0.0247      0.0040      0.0003
#> T>C_CTA     0.0062     0.0010 5.173842e-03     0.0154      0.0005      0.0001
#> T>C_CTC     0.0040     0.0008 1.093689e-02     0.0118      0.0019      0.0004
#> T>C_CTG     0.0270     0.0010 5.657774e-03     0.0154      0.0015      0.0001
#> T>C_CTT     0.0033     0.0045 1.112039e-02     0.0275      0.0020      0.0009
#> T>C_GTA     0.0122     0.0002 5.720779e-03     0.0091      0.0064      0.0010
#> T>C_GTC     0.0059     0.0001 5.181052e-03     0.0097      0.0101      0.0016
#> T>C_GTG     0.0115     0.0003 5.040707e-03     0.0094      0.0054      0.0019
#> T>C_GTT     0.0042     0.0081 8.488220e-03     0.0126      0.0124      0.0023
#> T>C_TTA     0.0059     0.0023 4.901072e-03     0.0156      0.0030      0.0002
#> T>C_TTC     0.0035     0.0018 6.081011e-03     0.0137      0.0097      0.0009
#> T>C_TTG     0.0106     0.0019 1.712485e-03     0.0098      0.0065      0.0007
#> T>C_TTT     0.0029     0.0024 1.000344e-02     0.0309      0.0099      0.0005
#> T>G_ATA     0.0000     0.0000 4.219719e-03     0.0335      0.0000      0.0000
#> T>G_ATC     0.0017     0.0000 7.936713e-04     0.0051      0.0040      0.0000
#> T>G_ATG     0.0007     0.0009 6.865800e-03     0.0052      0.0015      0.0001
#> T>G_ATT     0.0029     0.0000 2.939922e-03     0.0193      0.0137      0.0000
#> T>G_CTA     0.0001     0.0000 2.424051e-03     0.0216      0.0022      0.0000
#> T>G_CTC     0.0040     0.0008 2.096211e-03     0.0064      0.0018      0.0001
#> T>G_CTG     0.0050     0.0009 6.604046e-03     0.0126      0.0037      0.0011
#> T>G_CTT     0.0086     0.0013 4.866714e-03     0.0509      0.0182      0.0009
#> T>G_GTA     0.0000     0.0000 9.745787e-04     0.0072      0.0000      0.0000
#> T>G_GTC     0.0016     0.0000 5.248216e-04     0.0006      0.0020      0.0000
#> T>G_GTG     0.0010     0.0017 6.087753e-03     0.0050      0.0009      0.0010
#> T>G_GTT     0.0035     0.0009 5.427434e-03     0.0185      0.0030      0.0000
#> T>G_TTA     0.0009     0.0000 1.743221e-03     0.0502      0.0050      0.0000
#> T>G_TTC     0.0019     0.0010 2.549838e-03     0.0081      0.0092      0.0000
#> T>G_TTG     0.0011     0.0010 6.030395e-03     0.0088      0.0022      0.0003
#> T>G_TTT     0.0072     0.0014 7.223999e-03     0.0545      0.0633      0.0003
#>         Signature12  Signature13 Signature14 Signature15 Signature16
#> C>A_ACA      0.0077 3.347572e-04      0.0001      0.0013      0.0161
#> C>A_ACC      0.0047 6.487361e-04      0.0042      0.0040      0.0097
#> C>A_ACG      0.0017 3.814459e-05      0.0005      0.0000      0.0022
#> C>A_ACT      0.0046 8.466585e-04      0.0296      0.0057      0.0088
#> C>A_CCA      0.0135 1.710090e-03      0.0056      0.0106      0.0159
#> C>A_CCC      0.0112 1.159257e-03      0.0102      0.0084      0.0100
#> C>A_CCG      0.0028 2.441665e-04      0.0009      0.0015      0.0022
#> C>A_CCT      0.0071 1.256768e-03      0.1257      0.0228      0.0084
#> C>A_GCA      0.0062 1.321096e-04      0.0018      0.0024      0.0096
#> C>A_GCC      0.0056 7.542440e-04      0.0114      0.0099      0.0094
#> C>A_GCG      0.0015 9.682078e-05      0.0011      0.0013      0.0036
#> C>A_GCT      0.0027 5.511037e-04      0.0736      0.0309      0.0063
#> C>A_TCA      0.0066 4.816605e-02      0.0005      0.0057      0.0122
#> C>A_TCC      0.0099 1.732956e-02      0.0166      0.0062      0.0145
#> C>A_TCG      0.0019 2.293157e-03      0.0000      0.0025      0.0040
#> C>A_TCT      0.0049 1.863458e-02      0.0516      0.0145      0.0157
#> C>G_ACA      0.0031 3.775165e-03      0.0001      0.0011      0.0048
#> C>G_ACC      0.0015 9.208248e-04      0.0000      0.0001      0.0024
#> C>G_ACG      0.0000 1.989049e-05      0.0000      0.0006      0.0000
#> C>G_ACT      0.0025 3.860632e-03      0.0001      0.0010      0.0073
#> C>G_CCA      0.0019 5.333651e-03      0.0000      0.0009      0.0040
#> C>G_CCC      0.0019 9.123035e-04      0.0000      0.0003      0.0050
#> C>G_CCG      0.0000 2.854673e-04      0.0000      0.0004      0.0028
#> C>G_CCT      0.0019 6.383098e-03      0.0000      0.0017      0.0080
#> C>G_GCA      0.0012 1.349420e-03      0.0002      0.0022      0.0031
#> C>G_GCC      0.0011 7.888077e-04      0.0001      0.0005      0.0032
#> C>G_GCG      0.0000 8.179121e-05      0.0000      0.0006      0.0000
#> C>G_GCT      0.0009 2.680123e-03      0.0003      0.0018      0.0029
#> C>G_TCA      0.0024 2.802349e-01      0.0000      0.0009      0.0050
#> C>G_TCC      0.0027 6.388529e-02      0.0000      0.0026      0.0064
#> C>G_TCG      0.0000 9.528486e-03      0.0000      0.0000      0.0000
#> C>G_TCT      0.0015 3.302255e-01      0.0001      0.0015      0.0169
#> C>T_ACA      0.0121 1.480181e-03      0.0293      0.0117      0.0135
#> C>T_ACC      0.0054 4.050871e-05      0.0321      0.0169      0.0076
#> C>T_ACG      0.0019 3.165947e-04      0.0496      0.0309      0.0035
#> C>T_ACT      0.0067 0.000000e+00      0.0277      0.0066      0.0101
#> C>T_CCA      0.0187 3.647278e-03      0.0060      0.0051      0.0210
#> C>T_CCC      0.0094 1.004229e-03      0.0167      0.0088      0.0098
#> C>T_CCG      0.0046 1.292021e-04      0.0293      0.0155      0.0016
#> C>T_CCT      0.0205 6.117700e-04      0.0186      0.0096      0.0211
#> C>T_GCA      0.0124 2.934138e-03      0.0557      0.0476      0.0125
#> C>T_GCC      0.0092 2.515171e-03      0.0618      0.1345      0.0092
#> C>T_GCG      0.0023 3.205265e-03      0.0978      0.1829      0.0009
#> C>T_GCT      0.0115 1.687731e-03      0.0555      0.0800      0.0116
#> C>T_TCA      0.0138 1.138421e-01      0.0092      0.0112      0.0172
#> C>T_TCC      0.0095 1.502484e-02      0.0283      0.0156      0.0131
#> C>T_TCG      0.0010 6.102098e-03      0.0094      0.0104      0.0036
#> C>T_TCT      0.0090 2.814066e-02      0.0229      0.0076      0.0193
#> T>A_ATA      0.0058 1.243618e-03      0.0000      0.0021      0.0081
#> T>A_ATC      0.0065 1.060002e-03      0.0008      0.0090      0.0073
#> T>A_ATG      0.0046 1.025544e-04      0.0000      0.0009      0.0080
#> T>A_ATT      0.0060 9.718121e-04      0.0009      0.0079      0.0084
#> T>A_CTA      0.0017 2.851590e-04      0.0000      0.0016      0.0054
#> T>A_CTC      0.0086 2.222873e-04      0.0034      0.0043      0.0128
#> T>A_CTG      0.0031 2.869518e-04      0.0002      0.0027      0.0130
#> T>A_CTT      0.0054 8.004460e-07      0.0020      0.0007      0.0096
#> T>A_GTA      0.0015 1.379604e-04      0.0000      0.0000      0.0030
#> T>A_GTC      0.0039 2.528674e-04      0.0025      0.0038      0.0040
#> T>A_GTG      0.0019 1.535175e-04      0.0000      0.0026      0.0051
#> T>A_GTT      0.0031 1.227456e-04      0.0009      0.0043      0.0033
#> T>A_TTA      0.0057 4.032613e-04      0.0000      0.0008      0.0085
#> T>A_TTC      0.0132 7.718133e-04      0.0016      0.0019      0.0087
#> T>A_TTG      0.0037 2.451942e-04      0.0000      0.0006      0.0063
#> T>A_TTT      0.0094 1.603460e-04      0.0012      0.0003      0.0080
#> T>C_ATA      0.0466 5.494542e-04      0.0026      0.0115      0.0717
#> T>C_ATC      0.0317 1.455067e-04      0.0046      0.0073      0.0171
#> T>C_ATG      0.0487 2.791780e-04      0.0035      0.0073      0.0348
#> T>C_ATT      0.0250 2.674614e-04      0.0025      0.0063      0.0497
#> T>C_CTA      0.0312 4.275147e-05      0.0006      0.0032      0.0146
#> T>C_CTC      0.0385 4.330754e-05      0.0053      0.0050      0.0194
#> T>C_CTG      0.0415 5.427253e-05      0.0016      0.0128      0.0190
#> T>C_CTT      0.0324 7.639440e-07      0.0027      0.0035      0.0263
#> T>C_GTA      0.0573 1.054081e-03      0.0078      0.0195      0.0161
#> T>C_GTC      0.0274 5.298997e-04      0.0127      0.0499      0.0099
#> T>C_GTG      0.0410 6.300957e-04      0.0082      0.0154      0.0148
#> T>C_GTT      0.0405 6.761879e-04      0.0086      0.0094      0.0142
#> T>C_TTA      0.0500 7.840561e-04      0.0018      0.0068      0.0246
#> T>C_TTC      0.0342 1.329749e-04      0.0056      0.0099      0.0183
#> T>C_TTG      0.0320 3.740240e-05      0.0034      0.0039      0.0165
#> T>C_TTT      0.0371 1.797028e-04      0.0026      0.0047      0.0174
#> T>G_ATA      0.0012 3.861998e-04      0.0000      0.0000      0.0052
#> T>G_ATC      0.0010 1.123832e-04      0.0020      0.0006      0.0042
#> T>G_ATG      0.0018 3.600070e-05      0.0007      0.0012      0.0061
#> T>G_ATT      0.0030 1.924938e-04      0.0056      0.0029      0.0076
#> T>G_CTA      0.0013 1.509421e-05      0.0000      0.0000      0.0002
#> T>G_CTC      0.0049 1.476248e-04      0.0020      0.0019      0.0082
#> T>G_CTG      0.0045 4.598730e-05      0.0005      0.0007      0.0067
#> T>G_CTT      0.0063 4.637147e-04      0.0063      0.0045      0.0186
#> T>G_GTA      0.0000 1.848986e-05      0.0000      0.0000      0.0000
#> T>G_GTC      0.0004 9.337351e-05      0.0019      0.0022      0.0032
#> T>G_GTG      0.0011 1.057919e-05      0.0012      0.0037      0.0008
#> T>G_GTT      0.0032 5.374121e-05      0.0038      0.0190      0.0044
#> T>G_TTA      0.0019 5.465818e-04      0.0000      0.0000      0.0068
#> T>G_TTC      0.0027 2.353471e-04      0.0015      0.0004      0.0069
#> T>G_TTG      0.0011 4.790220e-07      0.0002      0.0009      0.0049
#> T>G_TTT      0.0032 6.705883e-04      0.0025      0.0033      0.0163
#>          Signature17  Signature18 Signature19  Signature20 Signature21
#> C>A_ACA 1.832019e-03 5.053642e-02      0.0107 1.179962e-03      0.0001
#> C>A_ACC 3.422356e-04 1.093982e-02      0.0074 2.211505e-03      0.0007
#> C>A_ACG 1.576225e-06 2.288073e-03      0.0005 1.616910e-07      0.0000
#> C>A_ACT 3.179665e-03 1.942409e-02      0.0074 3.008010e-03      0.0006
#> C>A_CCA 1.032430e-03 8.876811e-02      0.0112 1.737711e-02      0.0020
#> C>A_CCC 4.218801e-04 2.064139e-02      0.0159 3.650246e-02      0.0014
#> C>A_CCG 2.974628e-04 1.717840e-02      0.0018 1.248259e-02      0.0027
#> C>A_CCT 3.147943e-05 3.767696e-02      0.0096 1.034012e-01      0.0056
#> C>A_GCA 6.535405e-03 1.287242e-01      0.0032 1.116124e-03      0.0001
#> C>A_GCC 1.293804e-03 1.609288e-02      0.0031 3.271429e-03      0.0038
#> C>A_GCG 1.283663e-03 9.271417e-03      0.0000 3.751613e-03      0.0003
#> C>A_GCT 3.103767e-03 7.201707e-02      0.0072 1.293719e-02      0.0023
#> C>A_TCA 1.824558e-03 6.708064e-02      0.0051 6.882997e-04      0.0001
#> C>A_TCC 1.286518e-03 4.186901e-02      0.0063 4.058229e-03      0.0005
#> C>A_TCG 2.172942e-03 1.516238e-02      0.0018 1.001102e-03      0.0000
#> C>A_TCT 5.012316e-03 1.178778e-01      0.0087 1.222910e-02      0.0004
#> C>G_ACA 1.661704e-03 1.516888e-03      0.0058 6.970411e-04      0.0005
#> C>G_ACC 1.626898e-03 2.498785e-03      0.0019 2.059310e-03      0.0008
#> C>G_ACG 2.580190e-05 2.614509e-03      0.0000 1.273496e-05      0.0000
#> C>G_ACT 1.328508e-03 3.983011e-03      0.0072 8.485871e-04      0.0018
#> C>G_CCA 1.924979e-03 4.584638e-04      0.0002 1.331626e-05      0.0001
#> C>G_CCC 7.093678e-04 3.654228e-03      0.0028 5.481447e-04      0.0000
#> C>G_CCG 3.517869e-06 5.405774e-03      0.0002 0.000000e+00      0.0000
#> C>G_CCT 9.537763e-04 5.519605e-03      0.0008 1.207306e-04      0.0003
#> C>G_GCA 1.296324e-03 9.238797e-04      0.0020 1.339769e-03      0.0004
#> C>G_GCC 2.607609e-04 6.486479e-04      0.0000 1.811151e-03      0.0042
#> C>G_GCG 4.377552e-06 5.428400e-04      0.0000 6.375380e-07      0.0000
#> C>G_GCT 1.820393e-03 1.376725e-03      0.0000 1.238264e-03      0.0004
#> C>G_TCA 3.926847e-04 1.802969e-03      0.0005 2.429847e-05      0.0002
#> C>G_TCC 6.556024e-04 7.863007e-03      0.0011 1.519986e-04      0.0003
#> C>G_TCG 1.364154e-05 1.513223e-03      0.0000 0.000000e+00      0.0000
#> C>G_TCT 2.852380e-03 3.567102e-03      0.0105 6.830600e-04      0.0003
#> C>T_ACA 9.350488e-03 3.825624e-03      0.0221 3.285805e-02      0.0051
#> C>T_ACC 4.224289e-03 2.465214e-03      0.0203 2.219605e-02      0.0062
#> C>T_ACG 9.229213e-03 8.463957e-03      0.0368 3.457201e-02      0.0256
#> C>T_ACT 5.735961e-03 5.605740e-03      0.0266 1.881956e-02      0.0030
#> C>T_CCA 2.126521e-02 9.767511e-03      0.0438 1.114519e-02      0.0034
#> C>T_CCC 3.847235e-03 5.654828e-03      0.0753 1.815275e-02      0.0091
#> C>T_CCG 1.317611e-02 1.618322e-02      0.0238 4.438494e-02      0.0318
#> C>T_CCT 9.648025e-03 5.153319e-03      0.1159 1.384714e-02      0.0040
#> C>T_GCA 2.068755e-03 2.703069e-03      0.0374 5.679395e-02      0.0163
#> C>T_GCC 3.254834e-04 4.199308e-03      0.0614 5.442986e-02      0.0392
#> C>T_GCG 7.511101e-03 1.189736e-02      0.0278 4.592689e-03      0.0019
#> C>T_GCT 9.256972e-04 1.913381e-03      0.0917 4.958086e-02      0.0166
#> C>T_TCA 1.610078e-02 1.492339e-02      0.0218 1.071032e-02      0.0064
#> C>T_TCC 1.677998e-02 9.962744e-03      0.0330 8.175510e-03      0.0086
#> C>T_TCG 1.627210e-02 1.154972e-02      0.0090 2.316390e-02      0.0126
#> C>T_TCT 1.982742e-02 5.398014e-03      0.0267 9.540087e-03      0.0064
#> T>A_ATA 6.130025e-04 3.037211e-03      0.0011 5.744014e-04      0.0083
#> T>A_ATC 1.878587e-03 8.111672e-04      0.0053 1.971462e-03      0.0292
#> T>A_ATG 1.087350e-03 4.503207e-03      0.0045 1.905866e-04      0.0027
#> T>A_ATT 2.045494e-03 2.736523e-03      0.0033 7.421104e-03      0.0043
#> T>A_CTA 3.007924e-05 4.123464e-04      0.0043 4.504236e-05      0.0000
#> T>A_CTC 1.022375e-02 2.491411e-03      0.0053 4.389056e-03      0.0037
#> T>A_CTG 4.467036e-03 5.523046e-03      0.0021 1.127836e-03      0.0025
#> T>A_CTT 3.124488e-02 3.313713e-03      0.0022 2.385541e-03      0.0010
#> T>A_GTA 8.545689e-06 1.875434e-03      0.0026 1.631298e-04      0.0000
#> T>A_GTC 2.573467e-03 1.510310e-03      0.0000 3.799180e-03      0.0142
#> T>A_GTG 3.538823e-03 1.505369e-03      0.0016 8.283337e-04      0.0031
#> T>A_GTT 3.879010e-03 3.611025e-03      0.0052 1.900747e-03      0.0041
#> T>A_TTA 6.988123e-04 4.663278e-03      0.0039 9.245480e-04      0.0027
#> T>A_TTC 2.579359e-03 2.578156e-03      0.0048 7.082068e-04      0.0045
#> T>A_TTG 2.113364e-05 2.016738e-03      0.0052 9.700313e-05      0.0000
#> T>A_TTT 3.883131e-03 2.818679e-03      0.0085 2.313489e-03      0.0008
#> T>C_ATA 2.697314e-03 2.933054e-03      0.0145 1.639506e-02      0.0265
#> T>C_ATC 8.078150e-03 1.513285e-03      0.0048 1.925179e-02      0.0165
#> T>C_ATG 4.432871e-03 3.329669e-03      0.0135 5.478492e-02      0.0224
#> T>C_ATT 3.140370e-03 5.491318e-03      0.0160 7.341270e-03      0.0189
#> T>C_CTA 7.427000e-03 1.921693e-04      0.0066 1.486770e-02      0.0219
#> T>C_CTC 3.965700e-02 3.050193e-03      0.0070 1.835616e-02      0.0100
#> T>C_CTG 2.860970e-02 2.596464e-03      0.0040 7.415056e-02      0.0211
#> T>C_CTT 1.018697e-01 1.310029e-03      0.0053 1.100112e-02      0.0204
#> T>C_GTA 1.140133e-05 5.291265e-04      0.0039 2.469317e-03      0.1426
#> T>C_GTC 9.657713e-03 2.279131e-03      0.0030 1.565633e-03      0.0665
#> T>C_GTG 4.654129e-03 5.871057e-03      0.0060 2.734176e-02      0.0884
#> T>C_GTT 8.149235e-03 4.771566e-03      0.0099 1.498265e-03      0.1003
#> T>C_TTA 1.708587e-03 3.023036e-04      0.0052 5.734230e-03      0.0389
#> T>C_TTC 1.107415e-02 6.260791e-03      0.0020 1.123425e-02      0.0313
#> T>C_TTG 5.932003e-03 1.541265e-03      0.0057 2.376761e-02      0.0249
#> T>C_TTT 7.208218e-03 1.357797e-03      0.0038 5.395186e-03      0.0274
#> T>G_ATA 5.313925e-05 3.861382e-04      0.0014 1.765291e-06      0.0000
#> T>G_ATC 6.468661e-03 1.254982e-03      0.0000 3.813879e-04      0.0031
#> T>G_ATG 3.481549e-03 2.369216e-03      0.0010 6.587581e-04      0.0016
#> T>G_ATT 3.092638e-02 3.386029e-04      0.0017 5.321252e-04      0.0018
#> T>G_CTA 9.146979e-05 6.100255e-05      0.0000 2.326408e-05      0.0000
#> T>G_CTC 1.981429e-02 2.015697e-03      0.0011 8.476495e-03      0.0005
#> T>G_CTG 1.344984e-02 5.311967e-03      0.0013 1.538378e-02      0.0004
#> T>G_CTT 2.614566e-01 2.341629e-03      0.0019 2.185395e-03      0.0032
#> T>G_GTA 3.919382e-05 1.341532e-03      0.0000 3.579149e-06      0.0000
#> T>G_GTC 9.078462e-03 9.204310e-07      0.0000 2.830575e-03      0.0006
#> T>G_GTG 4.782775e-03 9.695000e-03      0.0043 2.733219e-03      0.0004
#> T>G_GTT 6.349776e-02 2.890554e-03      0.0025 3.558559e-03      0.0008
#> T>G_TTA 1.334106e-04 3.631840e-05      0.0032 2.691470e-07      0.0000
#> T>G_TTC 9.613345e-03 3.233838e-03      0.0018 3.772344e-04      0.0018
#> T>G_TTG 4.522462e-03 7.546018e-04      0.0011 5.154216e-04      0.0003
#> T>G_TTT 5.804041e-02 2.126442e-03      0.0013 6.156567e-04      0.0003
#>          Signature22  Signature23  Signature24  Signature25  Signature26
#> C>A_ACA 0.0015040704 4.533607e-04 2.864599e-02 0.0098967680 2.039773e-03
#> C>A_ACC 0.0024510110 3.668005e-04 2.021464e-02 0.0069989288 1.487162e-03
#> C>A_ACG 0.0000000000 0.000000e+00 2.047900e-02 0.0014484430 2.839456e-04
#> C>A_ACT 0.0009224525 0.000000e+00 2.460015e-02 0.0049665650 5.978656e-04
#> C>A_CCA 0.0045496929 1.647394e-04 6.355928e-02 0.0148329479 3.705850e-03
#> C>A_CCC 0.0037644739 7.368748e-04 3.375700e-02 0.0078221753 3.980723e-03
#> C>A_CCG 0.0009001633 1.639537e-04 2.242899e-02 0.0012769767 8.117420e-04
#> C>A_CCT 0.0044398462 7.227318e-04 2.008652e-02 0.0125636547 1.903843e-02
#> C>A_GCA 0.0012983702 3.499075e-04 5.467645e-02 0.0134652951 1.375312e-03
#> C>A_GCC 0.0018096222 5.683378e-04 1.099460e-01 0.0064682981 1.961639e-03
#> C>A_GCG 0.0003338381 1.494178e-04 4.652487e-02 0.0006581031 1.322966e-05
#> C>A_GCT 0.0005191482 0.000000e+00 5.042871e-02 0.0104730803 1.934996e-03
#> C>A_TCA 0.0040516346 2.071290e-03 3.145126e-02 0.0069381466 2.680062e-03
#> C>A_TCC 0.0027397709 6.964102e-04 6.253829e-02 0.0102466074 2.031644e-03
#> C>A_TCG 0.0005131511 9.939363e-05 1.280259e-02 0.0000000000 2.650717e-04
#> C>A_TCT 0.0036056509 2.074302e-04 3.029768e-02 0.0246257872 3.017187e-03
#> C>G_ACA 0.0005276440 0.000000e+00 1.199304e-02 0.0080325676 1.272881e-03
#> C>G_ACC 0.0000000000 0.000000e+00 8.425120e-03 0.0016359704 1.528195e-03
#> C>G_ACG 0.0000000000 0.000000e+00 0.000000e+00 0.0000000000 3.072463e-04
#> C>G_ACT 0.0002994547 0.000000e+00 3.881383e-03 0.0034284497 2.498252e-03
#> C>G_CCA 0.0005820177 1.990492e-04 1.347682e-02 0.0072007904 1.279106e-03
#> C>G_CCC 0.0008252996 2.837760e-04 1.610569e-02 0.0045984731 1.214977e-03
#> C>G_CCG 0.0003821146 0.000000e+00 7.974720e-04 0.0008237093 2.078326e-04
#> C>G_CCT 0.0017606459 1.957753e-04 1.038867e-02 0.0001963603 2.297026e-03
#> C>G_GCA 0.0000000000 0.000000e+00 4.945061e-03 0.0041696555 1.320607e-03
#> C>G_GCC 0.0000000000 1.375011e-04 9.369876e-03 0.0023928049 1.845501e-03
#> C>G_GCG 0.0002146959 0.000000e+00 5.625727e-03 0.0004181939 2.050608e-04
#> C>G_GCT 0.0005384388 0.000000e+00 3.055770e-03 0.0019940122 1.225533e-03
#> C>G_TCA 0.0002254906 2.306089e-04 1.022086e-02 0.0097987246 4.201897e-03
#> C>G_TCC 0.0006286950 1.657870e-04 7.578365e-03 0.0032665071 2.807762e-03
#> C>G_TCG 0.0003958080 0.000000e+00 1.592714e-03 0.0020574368 6.197000e-09
#> C>G_TCT 0.0002573751 4.389558e-04 6.100727e-03 0.0164009168 8.019383e-03
#> C>T_ACA 0.0031133897 1.976715e-02 6.315518e-03 0.0209876181 5.907232e-03
#> C>T_ACC 0.0015248603 8.188058e-02 3.844327e-03 0.0131412837 1.062646e-02
#> C>T_ACG 0.0036256412 9.248715e-03 3.110664e-03 0.0196598943 1.993032e-02
#> C>T_ACT 0.0000000000 3.320965e-02 3.420419e-03 0.0057285193 1.133453e-02
#> C>T_CCA 0.0077313495 6.057263e-02 6.035447e-03 0.0143339947 6.594445e-03
#> C>T_CCC 0.0071108505 1.863756e-01 0.000000e+00 0.0059366437 6.510632e-03
#> C>T_CCG 0.0055020310 1.514908e-02 2.924419e-03 0.0115593916 1.190494e-02
#> C>T_CCT 0.0041697772 1.560154e-01 3.971202e-03 0.0136325969 6.238556e-03
#> C>T_GCA 0.0019153709 5.328206e-02 2.367905e-02 0.0100381120 9.606515e-03
#> C>T_GCC 0.0032291335 1.629183e-01 1.591920e-02 0.0069538752 1.950743e-02
#> C>T_GCG 0.0033005989 1.304649e-02 1.264808e-02 0.0104683630 2.250272e-02
#> C>T_GCT 0.0011140596 1.075904e-01 1.812978e-02 0.0098238207 1.730747e-02
#> C>T_TCA 0.0020167217 1.455861e-02 2.657189e-02 0.0110775208 1.130329e-02
#> C>T_TCC 0.0019908342 4.351045e-02 1.205569e-02 0.0064487875 1.080753e-02
#> C>T_TCG 0.0014984731 2.752641e-03 5.363360e-03 0.0140834162 1.036433e-02
#> C>T_TCT 0.0011822265 2.432879e-02 5.532499e-03 0.0124185859 7.249145e-03
#> T>A_ATA 0.0496192297 0.000000e+00 1.166747e-03 0.0204305271 4.459340e-03
#> T>A_ATC 0.0116667405 0.000000e+00 1.517306e-05 0.0088925607 1.282224e-02
#> T>A_ATG 0.0679154542 2.374185e-04 6.332456e-03 0.0062419485 1.172344e-03
#> T>A_ATT 0.0145999176 0.000000e+00 4.967216e-06 0.0148508776 3.993138e-03
#> T>A_CTA 0.0868985455 5.945285e-04 2.766635e-03 0.0262963231 3.561423e-03
#> T>A_CTC 0.0537068468 0.000000e+00 2.694462e-03 0.0157998831 3.901611e-03
#> T>A_CTG 0.2132208666 1.731204e-04 6.911165e-03 0.0628606064 2.390304e-03
#> T>A_CTT 0.0618275075 4.578131e-04 2.915200e-03 0.0366223511 1.635923e-03
#> T>A_GTA 0.0537091457 0.000000e+00 2.225127e-03 0.0362178008 2.242887e-03
#> T>A_GTC 0.0094893967 1.832038e-04 4.853149e-03 0.0058471096 5.207287e-03
#> T>A_GTG 0.0502939024 0.000000e+00 1.957586e-04 0.0382411793 1.358122e-03
#> T>A_GTT 0.0106014572 0.000000e+00 0.000000e+00 0.0201679544 2.513113e-03
#> T>A_TTA 0.0408105028 0.000000e+00 0.000000e+00 0.0218303290 6.281453e-04
#> T>A_TTC 0.0255703958 0.000000e+00 7.994129e-03 0.0034056867 5.073754e-03
#> T>A_TTG 0.0698452181 2.315256e-04 1.120214e-02 0.0101915204 2.013839e-05
#> T>A_TTT 0.0235790618 0.000000e+00 2.705788e-04 0.0145275393 1.235538e-03
#> T>C_ATA 0.0168399309 1.068449e-03 4.726009e-03 0.0206632828 5.502865e-02
#> T>C_ATC 0.0011082624 2.215732e-04 6.279493e-04 0.0095024784 2.759460e-02
#> T>C_ATG 0.0049136164 4.988015e-04 2.723759e-03 0.0199114925 5.179104e-02
#> T>C_ATT 0.0026305239 3.132405e-04 7.122368e-03 0.0144883903 3.907224e-02
#> T>C_CTA 0.0040273463 0.000000e+00 1.848877e-03 0.0090280149 3.788938e-02
#> T>C_CTC 0.0013524439 0.000000e+00 4.453690e-03 0.0107716324 2.174073e-02
#> T>C_CTG 0.0027496661 3.207049e-04 0.000000e+00 0.0133506613 4.724010e-02
#> T>C_CTT 0.0023633535 0.000000e+00 1.597389e-03 0.0089610300 2.074062e-02
#> T>C_GTA 0.0042174541 2.511686e-04 1.269750e-03 0.0174898239 9.805282e-02
#> T>C_GTC 0.0000000000 5.382185e-04 6.710340e-03 0.0040456124 4.022636e-02
#> T>C_GTG 0.0014748845 0.000000e+00 2.776850e-04 0.0134022463 4.462070e-02
#> T>C_GTT 0.0015610427 9.957697e-04 4.238478e-03 0.0141596937 5.552807e-02
#> T>C_TTA 0.0048036697 0.000000e+00 1.682291e-03 0.0145150501 2.878968e-02
#> T>C_TTC 0.0009121575 2.306089e-04 2.344528e-03 0.0074060835 3.663949e-02
#> T>C_TTG 0.0017817357 2.395138e-04 4.812742e-04 0.0048925312 1.914369e-02
#> T>C_TTT 0.0000000000 0.000000e+00 3.581480e-04 0.0170147808 2.307169e-02
#> T>G_ATA 0.0016519986 0.000000e+00 0.000000e+00 0.0052816797 8.065343e-05
#> T>G_ATC 0.0000000000 0.000000e+00 1.463951e-03 0.0000000000 1.631712e-04
#> T>G_ATG 0.0010868728 0.000000e+00 8.476125e-04 0.0071364216 1.255080e-03
#> T>G_ATT 0.0000000000 0.000000e+00 1.075455e-03 0.0000000000 1.850383e-03
#> T>G_CTA 0.0011203565 0.000000e+00 0.000000e+00 0.0000000000 9.539626e-05
#> T>G_CTC 0.0010645836 1.901443e-04 2.185437e-03 0.0028864543 3.919455e-03
#> T>G_CTG 0.0041885681 0.000000e+00 1.139097e-04 0.0117694925 5.497357e-03
#> T>G_CTT 0.0016294096 2.108350e-04 2.533718e-03 0.0085508075 6.788719e-03
#> T>G_GTA 0.0000000000 0.000000e+00 0.000000e+00 0.0055462269 3.739323e-05
#> T>G_GTC 0.0000000000 0.000000e+00 1.073632e-03 0.0000000000 2.460705e-03
#> T>G_GTG 0.0018775892 1.871324e-04 2.924697e-03 0.0086852444 8.172016e-04
#> T>G_GTT 0.0000000000 0.000000e+00 8.250983e-04 0.0027685717 7.833561e-03
#> T>G_TTA 0.0016516988 0.000000e+00 0.000000e+00 0.0020814799 8.534935e-06
#> T>G_TTC 0.0000000000 0.000000e+00 0.000000e+00 0.0005789788 2.718510e-03
#> T>G_TTG 0.0025727520 2.475019e-04 1.360505e-03 0.0094291959 1.369161e-03
#> T>G_TTT 0.0000000000 0.000000e+00 6.951578e-05 0.0078696716 2.568077e-03
#>          Signature27  Signature28  Signature29 Signature30
#> C>A_ACA 5.205627e-03 1.397439e-03 0.0699819873   0.0000000
#> C>A_ACC 4.738227e-03 9.171877e-04 0.0551523572   0.0000000
#> C>A_ACG 7.826979e-04 0.000000e+00 0.0178469840   0.0019673
#> C>A_ACT 2.718242e-03 5.134100e-04 0.0268047160   0.0000000
#> C>A_CCA 5.065073e-03 1.168516e-03 0.0514102117   0.0000000
#> C>A_CCC 2.234153e-03 3.342918e-04 0.0258256508   0.0000000
#> C>A_CCG 2.663122e-04 5.365200e-05 0.0144961833   0.0022624
#> C>A_CCT 3.100570e-03 1.866719e-04 0.0403550741   0.0000000
#> C>A_GCA 1.075587e-02 2.136629e-03 0.0780466101   0.0088530
#> C>A_GCC 1.244234e-02 4.292084e-04 0.0734609147   0.0093449
#> C>A_GCG 1.629628e-03 0.000000e+00 0.0173832969   0.0008853
#> C>A_GCT 7.412355e-03 1.318507e-03 0.0596024721   0.0081645
#> C>A_TCA 7.079285e-03 7.759028e-04 0.0268944065   0.0000000
#> C>A_TCC 6.268801e-03 8.292925e-04 0.0316081370   0.0000000
#> C>A_TCG 1.446403e-03 0.000000e+00 0.0125004682   0.0016722
#> C>A_TCT 9.806278e-03 1.998074e-03 0.0412071764   0.0000000
#> C>G_ACA 1.324163e-03 2.551950e-04 0.0093019711   0.0000000
#> C>G_ACC 1.771084e-03 2.688990e-04 0.0034790489   0.0000000
#> C>G_ACG 0.000000e+00 0.000000e+00 0.0001544335   0.0048200
#> C>G_ACT 1.322720e-03 3.076980e-04 0.0039767899   0.0000000
#> C>G_CCA 4.730649e-03 1.033661e-04 0.0026188632   0.0000000
#> C>G_CCC 1.261645e-03 1.443108e-04 0.0011115123   0.0000000
#> C>G_CCG 3.608566e-04 0.000000e+00 0.0015200470   0.0032461
#> C>G_CCT 2.680262e-04 2.909995e-04 0.0001397507   0.0000000
#> C>G_GCA 1.418166e-03 8.159141e-05 0.0002657599   0.0073775
#> C>G_GCC 1.346717e-03 1.052190e-04 0.0009450129   0.0065906
#> C>G_GCG 0.000000e+00 0.000000e+00 0.0000000000   0.0015739
#> C>G_GCT 4.560686e-03 1.596794e-04 0.0047967447   0.0000000
#> C>G_TCA 2.870434e-03 5.213704e-04 0.0008204449   0.0000000
#> C>G_TCC 3.107968e-03 6.673259e-04 0.0004441193   0.0000000
#> C>G_TCG 1.281041e-05 0.000000e+00 0.0001432972   0.0019673
#> C>G_TCT 2.130588e-03 1.245780e-03 0.0013698690   0.0000000
#> C>T_ACA 1.372329e-02 5.433792e-03 0.0051972213   0.0651190
#> C>T_ACC 5.424306e-03 1.815415e-03 0.0089764700   0.0543970
#> C>T_ACG 6.706429e-03 5.432039e-03 0.0316389964   0.0204604
#> C>T_ACT 4.475614e-03 1.629213e-03 0.0047696666   0.0219359
#> C>T_CCA 1.743848e-02 4.391533e-03 0.0091264718   0.0694472
#> C>T_CCC 7.144870e-03 2.647661e-03 0.0070676030   0.0638403
#> C>T_CCG 7.503562e-03 2.284564e-03 0.0198751804   0.0173126
#> C>T_CCT 7.148028e-03 2.423984e-03 0.0016340587   0.0342318
#> C>T_GCA 1.421531e-02 3.384241e-03 0.0116387930   0.0485934
#> C>T_GCC 2.014365e-02 1.687414e-03 0.0037817517   0.0494787
#> C>T_GCG 9.368468e-03 2.189309e-03 0.0196843880   0.0157387
#> C>T_GCT 7.589626e-03 1.043994e-03 0.0057613368   0.0188865
#> C>T_TCA 1.875778e-02 7.622258e-03 0.0120765831   0.0849892
#> C>T_TCC 7.989184e-03 5.116619e-03 0.0000000000   0.0903010
#> C>T_TCG 4.609131e-03 1.918028e-03 0.0158301968   0.0151485
#> C>T_TCT 1.188093e-02 6.193159e-03 0.0134621373   0.0455440
#> T>A_ATA 1.430764e-01 9.640309e-04 0.0000000000   0.0075743
#> T>A_ATC 1.464085e-03 3.365668e-03 0.0030495957   0.0037380
#> T>A_ATG 1.226281e-03 1.785754e-03 0.0039099820   0.0065906
#> T>A_ATT 2.535198e-03 5.653089e-03 0.0051055274   0.0093447
#> T>A_CTA 3.306010e-01 2.566580e-03 0.0017654475   0.0053118
#> T>A_CTC 2.813599e-03 1.961090e-02 0.0003833591   0.0067873
#> T>A_CTG 4.396857e-03 8.833600e-03 0.0035568452   0.0082628
#> T>A_CTT 5.430892e-03 1.094261e-02 0.0022327555   0.0000000
#> T>A_GTA 1.098519e-01 5.514608e-04 0.0029300981   0.0037380
#> T>A_GTC 2.345568e-05 3.163988e-03 0.0000000000   0.0033445
#> T>A_GTG 1.565937e-03 2.152847e-03 0.0005336324   0.0050167
#> T>A_GTT 0.000000e+00 2.007733e-03 0.0000000000   0.0063939
#> T>A_TTA 5.281389e-02 2.032376e-04 0.0011912074   0.0076726
#> T>A_TTC 4.709179e-05 1.018096e-02 0.0016826640   0.0064922
#> T>A_TTG 0.000000e+00 3.250427e-03 0.0024387560   0.0036396
#> T>A_TTT 1.929771e-03 1.649281e-02 0.0000000000   0.0000000
#> T>C_ATA 5.047301e-03 4.092747e-03 0.0101191812   0.0000000
#> T>C_ATC 1.845691e-03 7.880729e-03 0.0063496827   0.0087547
#> T>C_ATG 5.688814e-03 5.520834e-03 0.0146596103   0.0090498
#> T>C_ATT 2.592755e-03 3.232484e-03 0.0078924134   0.0000000
#> T>C_CTA 4.899079e-03 2.073800e-02 0.0000000000   0.0000000
#> T>C_CTC 5.724448e-03 4.328380e-02 0.0027749537   0.0000000
#> T>C_CTG 5.661479e-03 3.708826e-02 0.0069024147   0.0000000
#> T>C_CTT 3.900589e-03 5.052716e-03 0.0030414249   0.0000000
#> T>C_GTA 4.777380e-03 3.655048e-03 0.0070908038   0.0000000
#> T>C_GTC 1.964954e-03 7.071175e-03 0.0071113275   0.0065906
#> T>C_GTG 1.229168e-03 8.232238e-03 0.0056430636   0.0078694
#> T>C_GTT 3.747225e-03 4.078845e-04 0.0112376743   0.0084596
#> T>C_TTA 1.721015e-03 9.929552e-03 0.0114986660   0.0000000
#> T>C_TTC 3.489213e-03 2.901056e-02 0.0035842150   0.0091481
#> T>C_TTG 1.152847e-03 1.568322e-02 0.0016942967   0.0060004
#> T>C_TTT 1.527957e-03 1.729156e-02 0.0037997058   0.0000000
#> T>G_ATA 1.136337e-03 6.364676e-03 0.0000000000   0.0053118
#> T>G_ATC 9.589764e-04 1.661091e-02 0.0000000000   0.0024592
#> T>G_ATG 6.766061e-05 6.592842e-03 0.0029251714   0.0062955
#> T>G_ATT 2.560277e-04 8.639290e-02 0.0000000000   0.0095416
#> T>G_CTA 0.000000e+00 9.589842e-03 0.0011168548   0.0046233
#> T>G_CTC 3.812450e-04 3.391322e-02 0.0000000000   0.0060004
#> T>G_CTG 4.032572e-04 1.851648e-02 0.0021696933   0.0073775
#> T>G_CTT 1.439096e-03 1.189671e-01 0.0011436633   0.0091481
#> T>G_GTA 0.000000e+00 0.000000e+00 0.0000000000   0.0032461
#> T>G_GTC 5.079056e-05 0.000000e+00 0.0000000000   0.0018690
#> T>G_GTG 4.742919e-03 4.217808e-03 0.0036405938   0.0033445
#> T>G_GTT 2.298476e-03 3.164900e-02 0.0061810238   0.0056069
#> T>G_TTA 1.189022e-03 9.389659e-03 0.0000000000   0.0086563
#> T>G_TTC 2.802954e-04 3.011708e-02 0.0000000000   0.0043282
#> T>G_TTG 2.353056e-03 1.269875e-02 0.0003536960   0.0082628
#> T>G_TTT 1.395613e-04 2.336598e-01 0.0061048341   0.0000000
#> 
#> Slot "exposures":
#>             TCGA-56-7582-01A-11D-2042-08 TCGA-77-7335-01A-11D-2042-08
#> Signature1                  3.848045e+01                 3.717409e+01
#> Signature2                  2.115756e-15                 2.608215e-15
#> Signature3                  2.685202e+01                 3.885659e+01
#> Signature4                  2.878543e+01                 6.479186e+01
#> Signature5                  2.398360e-15                 2.353065e-15
#> Signature6                  5.463688e-03                 1.652035e+01
#> Signature7                  1.191818e+01                 2.113318e+01
#> Signature8                  2.421394e-15                 2.165686e-15
#> Signature9                  2.195489e-15                 2.510472e-15
#> Signature10                 2.047161e-15                 1.972093e-15
#> Signature11                 2.192780e-15                 2.125077e-15
#> Signature12                 2.500387e-15                 2.387488e-15
#> Signature13                 4.855745e+00                 5.155692e+01
#> Signature14                 2.102671e-15                 1.883532e-15
#> Signature15                 2.300256e-15                 2.241178e-15
#> Signature16                 5.627097e-07                 5.778677e-07
#> Signature17                 2.067814e-15                 2.747507e-15
#> Signature18                 1.530343e+01                 1.281348e+00
#> Signature19                 2.530890e-15                 2.278516e-15
#> Signature20                 1.706963e+01                 8.996588e+00
#> Signature21                 3.426460e-01                 1.344370e+00
#> Signature22                 1.012177e+00                 6.931093e+00
#> Signature23                 2.570612e-15                 2.153941e-15
#> Signature24                 2.323363e+01                 1.031160e-02
#> Signature25                 2.320813e-15                 2.483758e-15
#> Signature26                 2.208179e-15                 2.429640e-15
#> Signature27                 1.956772e-15                 2.447203e-15
#> Signature28                 1.755932e-15                 2.718945e-15
#> Signature29                 2.426107e-15                 2.167524e-15
#> Signature30                 2.414120e+01                 2.540331e+01
#>             TCGA-94-7557-01A-11D-2122-08 TCGA-97-7938-01A-11D-2167-08
#> Signature1                  1.893511e+01                 5.995738e+00
#> Signature2                  1.999113e-15                 2.141450e-15
#> Signature3                  4.024637e+00                 1.807293e+00
#> Signature4                  4.864076e+00                 3.861682e+01
#> Signature5                  2.438016e-15                 2.016099e-15
#> Signature6                  6.759240e+00                 5.102840e-03
#> Signature7                  2.004344e+01                 8.794115e+00
#> Signature8                  2.328546e-15                 2.274807e-15
#> Signature9                  2.106516e-15                 1.913530e-15
#> Signature10                 2.526891e-15                 2.518896e-15
#> Signature11                 2.631768e-15                 2.206736e-15
#> Signature12                 2.373845e-15                 2.031406e-15
#> Signature13                 3.189745e+01                 1.305966e+01
#> Signature14                 2.364401e-15                 2.376667e-15
#> Signature15                 3.093584e-15                 2.486528e-15
#> Signature16                 1.250481e+01                 4.677128e-07
#> Signature17                 1.288641e-15                 1.413866e-15
#> Signature18                 4.101353e+00                 1.893098e+01
#> Signature19                 2.868640e-15                 1.955754e-15
#> Signature20                 6.135323e+00                 7.195111e+00
#> Signature21                 2.169415e+00                 5.794154e-01
#> Signature22                 1.556695e-01                 4.698027e+00
#> Signature23                 3.085810e-15                 2.012678e-15
#> Signature24                 8.630665e-03                 1.203777e+01
#> Signature25                 2.116190e-15                 1.970347e-15
#> Signature26                 2.628320e-15                 2.139247e-15
#> Signature27                 2.009951e-15                 2.011775e-15
#> Signature28                 2.356250e-15                 2.391588e-15
#> Signature29                 2.248478e-15                 2.373682e-15
#> Signature30                 5.400848e+00                 5.279966e+00
#>             TCGA-EE-A3J5-06A-11D-A20D-08 TCGA-ER-A197-06A-32D-A197-08
#> Signature1                  2.961999e+00                 1.479083e+00
#> Signature2                  1.564971e-15                 3.051840e-16
#> Signature3                  4.498477e-01                 2.083453e-01
#> Signature4                  7.179628e-01                 7.607862e-01
#> Signature5                  2.126433e-15                 1.275245e-15
#> Signature6                  2.955223e-03                 3.390919e-03
#> Signature7                  1.010365e+02                 1.958114e+00
#> Signature8                  2.005685e-15                 1.436914e-15
#> Signature9                  3.337025e-15                 8.234475e-16
#> Signature10                 1.131941e-15                 4.845864e-16
#> Signature11                 2.195293e-15                 2.169710e-15
#> Signature12                 2.451459e-15                 1.570029e-15
#> Signature13                 1.368678e-02                 3.457752e-03
#> Signature14                 9.753197e-16                 1.535067e-15
#> Signature15                 1.242437e-15                 2.191223e-15
#> Signature16                 5.352687e-07                 2.972437e-07
#> Signature17                 8.374366e-16                 5.372505e-16
#> Signature18                 9.613142e-01                 3.369291e-01
#> Signature19                 2.081140e-15                 2.066190e-15
#> Signature20                 3.024916e-01                 2.984900e-01
#> Signature21                 6.665302e+00                 3.320641e+00
#> Signature22                 3.235105e+00                 1.306803e+00
#> Signature23                 2.383550e-15                 2.389548e-15
#> Signature24                 4.664651e-03                 5.837055e-03
#> Signature25                 2.507415e-15                 1.248534e-15
#> Signature26                 2.275095e-15                 1.752700e-15
#> Signature27                 2.147240e-15                 4.508983e-16
#> Signature28                 1.651640e-15                 6.202035e-16
#> Signature29                 1.760761e-15                 1.366590e-15
#> Signature30                 5.648123e+00                 1.318122e+00
#>             TCGA-ER-A19O-06A-11D-A197-08
#> Signature1                  4.096906e+00
#> Signature2                  1.634428e-15
#> Signature3                  1.054276e+00
#> Signature4                  3.843626e-01
#> Signature5                  2.092936e-15
#> Signature6                  4.388855e-03
#> Signature7                  4.027624e+01
#> Signature8                  1.555093e-15
#> Signature9                  1.491339e-15
#> Signature10                 2.151726e-15
#> Signature11                 1.960235e-15
#> Signature12                 1.481362e-15
#> Signature13                 1.489710e-02
#> Signature14                 1.450357e-15
#> Signature15                 2.246871e-15
#> Signature16                 4.241981e-07
#> Signature17                 1.381962e-15
#> Signature18                 7.266010e-02
#> Signature19                 1.869588e-15
#> Signature20                 2.591193e-01
#> Signature21                 3.160216e-01
#> Signature22                 7.113510e-01
#> Signature23                 2.134947e-15
#> Signature24                 4.090909e-03
#> Signature25                 2.282894e-15
#> Signature26                 1.621136e-15
#> Signature27                 4.155933e-15
#> Signature28                 2.247465e-15
#> Signature29                 9.678221e-16
#> Signature30                 3.805689e+00
#> 
#> Slot "table_name":
#> [1] "SBS96"
#> 
#> Slot "algorithm":
#> [1] "posterior_LDA"
#> 
#> Slot "musica":
#> An object of class "musica"
#> Slot "variants":
#>        chr     start       end ref alt                       sample
#>   1:  chr1  11020563  11020563   C   A TCGA-94-7557-01A-11D-2122-08
#>   2:  chr1  43430030  43430030   G   T TCGA-94-7557-01A-11D-2122-08
#>   3:  chr1  58682403  58682403   A   G TCGA-94-7557-01A-11D-2122-08
#>   4:  chr1 109508295 109508295   C   G TCGA-94-7557-01A-11D-2122-08
#>   5:  chr1 156384826 156384826   A   C TCGA-94-7557-01A-11D-2122-08
#>  ---                                                               
#> 907: chr19  54885292  54885292   G   A TCGA-ER-A19O-06A-11D-A197-08
#> 908: chr20  49374328  49374328   A   T TCGA-ER-A19O-06A-11D-A197-08
#> 909:  chrX  73213768  73213768   T   G TCGA-ER-A19O-06A-11D-A197-08
#> 910:  chrX 101292834 101292834   G   A TCGA-ER-A19O-06A-11D-A197-08
#> 911:  chrX 107526690 107526690   C   T TCGA-ER-A19O-06A-11D-A197-08
#>      Variant_Type
#>   1:          SBS
#>   2:          SBS
#>   3:          SBS
#>   4:          SBS
#>   5:          SBS
#>  ---             
#> 907:          SBS
#> 908:          SBS
#> 909:          SBS
#> 910:          SBS
#> 911:          SBS
#> 
#> Slot "count_tables":
#> $SBS96
#> Count_Table:  SBS96 
#> Motifs: 96 
#> Samples: 7 
#>  
#> **Annotations: 
#>            motif mutation context
#> C>A_ACA C>A_ACA      C>A     ACA
#> C>A_ACC C>A_ACC      C>A     ACC
#> C>A_ACG C>A_ACG      C>A     ACG
#> C>A_ACT C>A_ACT      C>A     ACT
#> C>A_CCA C>A_CCA      C>A     CCA
#> C>A_CCC C>A_CCC      C>A     CCC
#> 7           ...      ...     ... 
#> 
#> **Features: 
#>    mutation
#> 1  C>A_TCA
#> 2  C>A_ACT
#> 3  T>C_ATT
#> 4  C>G_TCT
#> 5  T>G_CTC
#> 6  C>G_CCC
#> 7      ... 
#> 
#> **Types: 
#>  SBS
#>  
#> **Color Variable: 
#>  mutation
#>  
#> **Color Mapping: 
#>  #5ABCEBFF
#>  #050708FF
#>  #D33C32FF
#>  #CBCACBFF
#>  #ABCD72FF
#>  #E7C9C6FF
#>  
#> **Descriptions: 
#>  Single Base Substitution table with one base upstream and downstream
#> 
#> $SBS96
#> Count_Table:  SBS96 
#> Motifs: 96 
#> Samples: 7 
#>  
#> **Annotations: 
#>            motif mutation context
#> C>A_ACA C>A_ACA      C>A     ACA
#> C>A_ACC C>A_ACC      C>A     ACC
#> C>A_ACG C>A_ACG      C>A     ACG
#> C>A_ACT C>A_ACT      C>A     ACT
#> C>A_CCA C>A_CCA      C>A     CCA
#> C>A_CCC C>A_CCC      C>A     CCC
#> 7           ...      ...     ... 
#> 
#> **Features: 
#>    mutation
#> 1  C>A_CCG
#> 2  C>A_GCA
#> 3  T>C_ATG
#> 4  C>G_TGC
#> 5  T>G_AGC
#> 6  C>G_CCT
#> 7      ... 
#> 
#> **Types: 
#>  SBS
#>  
#> **Color Variable: 
#>  mutation
#>  
#> **Color Mapping: 
#>  #5ABCEBFF
#>  #050708FF
#>  #D33C32FF
#>  #CBCACBFF
#>  #ABCD72FF
#>  #E7C9C6FF
#>  
#> **Descriptions: 
#>  Single Base Substitution table with one base upstream and downstream
#> 
#> 
#> Slot "sample_annotations":
#>                         Samples Tumor_Subtypes
#> 1: TCGA-94-7557-01A-11D-2122-08           Lung
#> 2: TCGA-56-7582-01A-11D-2042-08           Lung
#> 3: TCGA-77-7335-01A-11D-2042-08           Lung
#> 4: TCGA-97-7938-01A-11D-2167-08           Lung
#> 5: TCGA-EE-A3J5-06A-11D-A20D-08         Breast
#> 6: TCGA-ER-A197-06A-32D-A197-08         Breast
#> 7: TCGA-ER-A19O-06A-11D-A197-08         Breast
#> 
#> 
#> Slot "umap":
#> <0 x 0 matrix>
#>